% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{add_phiprime}
\alias{add_phiprime}
\title{add phiprime contours to Linf_K scatterplot}
\usage{
add_phiprime(gridsize = 20, ...)
}
\arguments{
\item{gridsize}{default = 20}

\item{...}{Additional arguments passed to `par`.}
}
\value{
plot
}
\description{
add phiprime contours to Linf_K scatterplot
}
