% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBB.R
\name{AG}
\alias{AG}
\title{Aggregation function for LBB}
\usage{
AG(dat)
}
\arguments{
\item{dat}{Data.frame with columns: Year, Length (in cm), and CatchNo}
}
\value{
A data.frame with columns Length and Freq
}
\description{
Function to aggregate data by year
}
\references{
R. Froese, H. Winker, G. Coro, N. Demirel, A.C. Tsikliras, D. Dimarchopoulou,
G. Scarcella, W.N. Probst, M. Dureuil, and D. Pauly (2018) A new approach
for estimating stock status from length frequency data. ICES Journal of Marine Science. DOI: 10.1093/icesjms/fsy078
}
