% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBB.R
\name{BH}
\alias{BH}
\title{B & H equations for LBB}
\usage{
BH(AllLength, Linf, MK, FK, GausSel, selpar1, selpar2)
}
\arguments{
\item{AllLength}{Vector with all lengths}

\item{Linf}{Asymptotic length}

\item{MK}{MK ratio}

\item{FK}{FK ratio}

\item{GausSel}{Logical; indicating the selectivity pattern. If FALSE (default) trawl-like,
if TRUE gaussian selectivity is assumed.}

\item{selpar1}{Selectivity parameter 1}

\item{selpar2}{Selectivity parameter 2}
}
\value{
A list with B/B0 and Y/R
}
\description{
Exploited B/B0 ratio from B&H equations for variable F
}
\details{
Assuming that observed lengths are the lower bounds of length
    classes get lowest exploited (>= 0.01 F) length class and class width.
}
\references{
R. Froese, H. Winker, G. Coro, N. Demirel, A.C. Tsikliras, D. Dimarchopoulou,
G. Scarcella, W.N. Probst, M. Dureuil, and D. Pauly (2018) A new approach
for estimating stock status from length frequency data. ICES Journal of Marine Science. DOI: 10.1093/icesjms/fsy078
}
