% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bhattacharya.R
\name{Bhattacharya}
\alias{Bhattacharya}
\title{Bhattacharya's method}
\usage{
Bhattacharya(param)
}
\arguments{
\item{param}{A list consisting of following parameters:
\code{$age} or \code{$midLengths} midpoints of the length class as vector (length frequency
data) or ages as vector (age composition data),
\code{catch} Catch as vector, or a matrix with catches of subsequent years if
the catch curve with constat time intervals should be applied;}
}
\description{
Separating normal distributions of several cohorts in distinct
  distributions representing cohorts. Second sentence.
}
\details{
Bhattacharya
}
\examples{
\donttest{
 data(synLFQ1)
 Bhattacharya(synLFQ1)
}
}
\references{
Bhattacharya, C.G., 1967. A simple method of resolution of a distribution into Gaussian
components, \emph{Biometrics}, 23:115-135

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

