% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tnew.R
\name{calc_tnew}
\alias{calc_tnew}
\title{calculate time for new length}
\usage{
calc_tnew(par = NULL, Ltnew = NULL, Lt = NULL, t = NULL,
  tincr = 0.01)
}
\arguments{
\item{par}{VBGF parameters. If soVBGF, then ts should relate
to time of calendar year rather than time of age year.}

\item{Ltnew}{New length for which time is calculated}

\item{Lt}{Known length at time t}

\item{t}{known time t}

\item{tincr}{time increment (year fraction) used for iterative
hind-/fore-casting. Use smaller values for increased precision
(Default: tincr = 0.01)}
}
\value{
numeric value representing year decimal
}
\description{
calculate time for new length
}
\examples{

data(synLFQ4)

# time at length = 0
calc_tnew(par = synLFQ4$par, Lt = 70, t = 2000, Ltnew = 0)

# time at length = 75
calc_tnew(par = synLFQ4$par, Lt = 70, t = 2000, Ltnew = 75)

}
