% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trecr.R
\name{calc_trecr}
\alias{calc_trecr}
\title{Calculate time at recruitment}
\usage{
calc_trecr(par = NULL, Lrecr = 0, Lt = NULL, t = NULL, tincr = 0.01)
}
\arguments{
\item{par}{\code{\link{VBGF}} parameters}

\item{Lrecr}{Length at recruitment. This can be e.g. the length at recruitment to the
population or fishery}

\item{Lt}{known length at time \code{t}}

\item{t}{time (class is numeric or date)}

\item{tincr}{resolution to use for hindcasting time at recruitment (default: 0.01)}
}
\value{
a numeric value (typically decimal year)
}
\description{
The \code{calc_trecr} function calculates the time at recruitment based
on \code{\link{VBGF}} parameters, a known length at time \code{t}
(i.e. date or numeric class) (\code{Lt}), and a defined length at
recruitment \code{Lrecr}. The solution is based on a iterative hindcast of
the soVBGF. \code{calc_trecr} is called internally by
\code{\link{recruitment2}}
}
\examples{
lfq <- synLFQ4
lfq$par <- list(
  Linf = 80, K = 0.5, t_anchor = 0.25, C = 0.75, ts = 0.5
)
age <- 1
Lt <- VBGF(param = lfq$par, t = age)
trecr <- calc_trecr(par = lfq$par, Lrecr = 0, Lt = Lt, t = age)
trecr
lfq$par$t_anchor # should be equivalent


}
