% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLt.R
\name{calcLt}
\alias{calcLt}
\title{Calculation of lengths of multiple cohorts}
\usage{
calcLt(lfq, par = list(Linf = 100, K = 0.1, t_anchor = 0.25, C = 0, ts = 0),
  agemax = NULL, flagging.out = TRUE, lty = 2, lwd = 1, col = 1,
  draw = FALSE, tincr = 0.05)
}
\arguments{
\item{lfq}{a list of the class "lfq" consisting of following parameters:
\itemize{
  \item \strong{midLengths} midpoints of the length classes,
  \item \strong{dates} dates of sampling times (class Date),
  \item \strong{catch} matrix with catches/counts per length class (row) and sampling date (column),
  \item \strong{samplingPeriod} length of sampling period in years,
  \item \strong{samplingDays} time of sampling times in relation to first sampling time,
  \item \strong{delta_t} array with time differences between relative sampling time set to zero and
     other sampling times,
  \item \strong{rcounts} restructured frequencies,
  \item \strong{peaks_mat} matrix with positive peaks with distinct values,
  \item \strong{ASP} available sum of peaks, sum of posititve peaks which could be potential be hit by
     growth curves;
}}

\item{par}{a list with following growth parameters:
 \itemize{
  \item \strong{Linf} length infinity in cm (default: 100),
  \item \strong{K} curving coefficient (default: 0.1),
  \item \strong{t_anchor} time point anchoring growth curves in year-length
  coordinate system, corrsponds to peak spawning month (range: 0 to 1, default: 0.25),
  \item \strong{C} amplitude of growth oscillation (range: 0 to 1, default: 0),
  \item \strong{ts} summer point (ts = WP - 0.5) (range: 0 to 1, default: 0);
}}

\item{agemax}{maximum age of species; default NULL, then estimated from Linf}

\item{flagging.out}{logical; should positive peaks be flagged out?}

\item{lty}{The line type. Line types can either be specified as an integer
(0=blank, 1=solid, 2=dashed (default), 3=dotted, 4=dotdash, 5=longdash,
6=twodash) or as one of the character strings "blank", "solid", "dashed",
"dotted", "dotdash", "longdash", or "twodash", where "blank" uses 'invisible
lines' (i.e., does not draw them).}

\item{lwd}{The line width, a positive number, defaulting to 2. The interpretation
is device-specific, and some devices do not implement line widths less
than one. (See the help on the device for details of the interpretation.)}

\item{col}{A specification for the default plotting color. See section
'Color Specification'.}

\item{draw}{logical; indicating whether growth curves should be added to
existing lfq plot}

\item{tincr}{step for plotting}
}
\value{
A list with the input parameters and following list objects:
\itemize{
  \item \strong{samplingPeriod}: length of sampling period in years,
  \item \strong{samplingDays}: time of sampling times in relation to first sampling time,
  \item \strong{delta_t}: array with time differences between relative sampling time set to zero and
     other sampling times,
  \item \strong{rcounts}: restructured frequencies,
  \item \strong{peaks_mat}: matrix with positive peaks with distinct values,
  \item \strong{startingPoints}: starting sample and starting length yielding in best fit
  \item \strong{agemax}: maximum age assumed
}
}
\description{
calculates growth curves for set of growth parameters
}
\details{
Anchoring time point t_anchor 0.25 for example peak spawning in April or something
}
\examples{
data(trout)
res <- lfqRestructure(trout)
plot(res)
calcLt(res,par=list(Linf=16,K=0.8,t_anchor=0.28),draw=TRUE)

}
\references{
Pauly, D. and N. David, 1981. ELEFAN I, a BASIC program for the objective extraction of
growth parameters from length-frequency data. \emph{Meeresforschung}, 28(4):205-211
}

