% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchCurve.R
\name{catchCurve}
\alias{catchCurve}
\title{Catch curve}
\usage{
catchCurve(param, catch_column = NA, cumulative = FALSE,
  calc_ogive = FALSE)
}
\arguments{
\item{param}{a list consisting of following parameters:
\itemize{
  \item \strong{age} or \strong{midLengths} midpoints of the length class as vector (length frequency
  data) or ages as vector (age composition data),
  \item \strong{Linf} Infinite length for investigated species in cm [cm],
  \item \strong{K} Growth coefficent for investigated species per year [1/year],
  \item \strong{t0} Theoretical time zero, at which individuals of this species hatch,
  \item \strong{catch} Catch as vector, or a matrix with catches of subsequent years if
  the catch curve with constat time intervals should be applied;
}}

\item{catch_column}{numerical; indicating the column of the catch matrix which should be
used for the analysis.}

\item{cumulative}{logical; if \code{TRUE} instead of normal catch curve the cumulative
catch curve is applied (Jones and van Zalinge method)}

\item{calc_ogive}{logical; if \code{TRUE} the selection ogive is additionally
calculated from the catch curve}
}
\description{
This is a function to calculate the total mortality (Z) from length
  composition data via the length converted catch curve or from age at length data
  with catch curve. It has the option to estimate gear selectivity based on the catch curve.
}
\details{
For variable parameter system vectors are reuqired for constant parameter
  systems matrices or data.frames have to be inserted. or vectors The length converted
  linearised catch curve is used to calculate the total mortality (Z). This function
  includes a so called locator function, which asks you to choose points from a graph
  manually. Based on these points the regression line is calculated. When selection ogive
  is calculated by means of the catch curve the assumption is made, that Z is constant
  for all year classes or length groups respectively. Accoring to Sparre and Venema
  (1998) this assumption might be true, because F is smaller for young fish
  (Selectivity) while M is higher for young fish (high natural mortality). The selectivity
  for not fully exploited old fish (e.g. due to gillnet fishery) can not be calculated yet
  by use of the catch curve.
}
\examples{
\donttest{
#_______________________________________________
# Variable paramter system (with catch vector)
# based on length frequency data
# load data
data(goatfish)

# run model
catchCurve(goatfish)

# based on age composition data
# load data
data(whiting)

# run model
catchCurve(whiting, catch_column = 1)

#_______________________________________________
# Constant parameter system based on age composition data (with catch matrix)
catchCurve(whiting)


#_______________________________________________
# Cumulative Catch Curve
# based on length frequency data
# load data
data(goatfish)

# run model
catchCurve(goatfish, cumulative = TRUE)


# based on age composition data
data(synCAA2)

# run model
catchCurve(synCAA2, cumulative = TRUE)


#_______________________________________________
# Catch Curve with estimation of selection ogive
# load data
data(synLFQ3)

# run model
catchCurve(synLFQ3, calc_ogive = TRUE)

 }

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.

ICES, 1981. Report of the \emph{Ad hoc} working group on the use of effort data in
assessment, Copenhagen, 2-6 March 1981. ICES C.M. 1981/G:5 (mimeo)
}

