% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchCurveCum.R
\name{CatchCurveCum}
\alias{CatchCurveCum}
\title{Cumulative catch curve (The Jones and van Zalinge method)}
\usage{
CatchCurveCum(param)
}
\arguments{
\item{param}{A list consisting of following parameters:
\code{$age} or \code{$midLengths} midpoints of the length class as vector (length frequency
data) or ages as vector (age composition data),
\code{$Linf} Infinite length for investigated species in cm [cm],
\code{$K} Growth coefficent for investigated species per year [1/year],
\code{t0} Theoretical time zero, at which individuals of this species hatch,
\code{catch} Catch as vector, or a matrix with catches of subsequent years if
the catch curve with constat time intervals should be applied;}

\item{catch_column}{A number indicating the column of the catch matrix which should be
used for the anaalysis.}
}
\description{
This is a function to calculate the total mortality (Z) from length
   composition data via the length converted cumulative catch curve or from age
   composition data via the cumulative catch curve.
}
\details{
Not good for methods where the largest length groups are not represented fully (highly selective gear). The length converted linearised cumulative catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
\donttest{

# Cumulative catch curve based on length frequency data
# load data
data(goatfish)

# calculate cumulative catch and add to data list
goatfish$cumCatch <- rev(cumsum(rev(goatfish$catch)))

# run model
CatchCurveCum(goatfish)

# based on age composition data
data(synCAA2)

# calculate cumulative catch
synCAA2$cumCatch <- rev(cumsum(rev(synCAA2$catch)))

# run model
CatchCurveCum(synCAA2)

}

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

