% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchCurvePlusSelectOgive.R
\name{CatchCurvePlusSelectOgive}
\alias{CatchCurvePlusSelectOgive}
\title{Catch curve with selection ogive}
\usage{
CatchCurvePlusSelectOgive(param, catch_column = NA)
}
\arguments{
\item{param}{A list consisting of following parameters:
\code{$age} or \code{$midLengths} midpoints of the length class as vector (length frequency
data) or ages as vector (age composition data),
\code{$Linf} Infinite length for investigated species in cm [cm],
\code{$K} Growth coefficent for investigated species per year [1/year],
\code{t0} Theoretical time zero, at which individuals of this species hatch,
\code{catch} Catch as vector, or a matrix with catches of subsequent years if
the catch curve with constat time intervals should be applied;}

\item{catch_column}{A number indicating the column of the catch matrix which should be
used for the anaalysis.}
}
\description{
This is a function to calculate the total mortality (Z) from length
   composition data via the length converted catch curve or from age at length data
   with catch curve.
}
\details{
For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
\donttest{
# Variable paramter system
# based on length frequency data
data(goatfish)

param = goatfish
param = whiting

CatchCurvePlusSelectOgive(goatfish)


# based on age composition data
data(whiting)
CatchCurvePlusSelectOgive(whiting,catch_column = 2)

# Constant parameter system based on age composition data
CatchCurvePlusSelectOgive(whiting)
 }

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.

ICES, 1981. Report of the \emph{Ad hoc} working group on the use of effort data in
assessment, Copenhagen, 2-6 March 1981. ICES C.M. 1981/G:5 (mimeo)
}

