% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CohortAnalysis.R
\name{CohortAnalysis}
\alias{CohortAnalysis}
\title{Cohort analysis}
\usage{
CohortAnalysis(midLengths, catch, Linf, K, t0 = 0, M, terminalF, a, b,
  catchCorFac = NA)
}
\arguments{
\item{midLengths}{Midpoints of the length class as vector}

\item{catch}{Catch per sampling time as matrix or the total catch as vector.}

\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficient for investigated species per year [1/year].}

\item{t0}{Theoretical time zero, at which individuals of this species hatch.}

\item{M}{Natural mortality [1/year]}

\item{terminalF}{terminal fishing mortality}

\item{a}{length-weight relationship coefficent (W = a * L^b)}

\item{b}{length-weight relationship coefficent (W = a * L^b)}

\item{catchCorFac}{optional: Correction factor for catch, in case provided catch does spatially or temporarily not reflect catch for fishing ground of a whole year.}
}
\description{
Cohort analysis
}
\details{
Cohort analysis
}
\examples{
data("ex.CohortAnalysis")
output = CohortAnalysis(midLengths = ex.CohortAnalysis$midLengths, catch = ex.CohortAnalysis$catch, Linf = 130, K = 0.1, M = 0.28, terminalF = 0.28, a = 0.00001, b = 3)
output
}
\references{
Jones ???  Sparre?
}

