% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_GillnetSelect.R
\docType{data}
\name{data_GillnetSelect}
\alias{data_GillnetSelect}
\title{Example dataset for the use of \code{\link[FishTropR]{GillnetSelect}}}
\format{A list consiting of: the midlengths of size classes (\code{$midLengths}),
the number of fish caught with net 1 \code{$numNet1}, the number of fish caught with
net 2 (\code{$numNet2}), and the meshsizes of both nets (\code{$msNet1} & \code{$msNet2}).}
\source{
Garrod, D.J., 1961. The selection characteristics of nylon gill nets for \textit{Tilapia esculenta} Graham. J.Cons.CIEM, 26(2):191-203
}
\usage{
data(data_GillnetSelect)
}
\description{
bla bla

\itemize{
  \item bla1. Dateframe of lon/lat coordinates corresponding to columns of sst$field (5 deg resolution)
  \item bla2. Vector of monthly date values corresponding to rows of sst$field
  \item bla3. Matrix of sea level pressure values by month and lon/lat position.
}
}
\examples{
data(data_GillnetSelect)
summary(data_GillnetSelect)
str(data_GillnetSelect)

}
\keyword{dataset}
\keyword{selectivity}

