% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Predict_ThompsonBell.R
\docType{data}
\name{data_Predict_ThompsonBell}
\alias{data_Predict_ThompsonBell}
\title{Example dataset of length for use in Predict_ThompsonBell}
\format{A list consisting of: 1. a dataframe for lon/lat positions, 2. a vector of
monthly date values, and 3. a matrix of sst anomaly values by month and lon/lat position
(1906 rows, 264 columns)}
\usage{
data(data_Predict_ThompsonBell)
}
\description{
First sentence. second sentence.

\itemize{
  \item bla1. Dateframe of lon/lat coordinates corresponding to columns of sst$field (5 deg resolution)
  \item bla2. Vector of monthly date values corresponding to rows of sst$field
  \item bla3. Matrix of sea level pressure values by month and lon/lat position.
}
}
\examples{
data(data_Predict_ThompsonBell)
head(data_Predict_ThompsonBell)


}
\keyword{datasets}
\keyword{length}

