% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TrawlSelect.R
\docType{data}
\name{data_TrawlSelect}
\alias{data_TrawlSelect}
\title{Example dataset of length for use in TrawlSelect}
\format{A list consisting of: 1. a da}
\usage{
data(data_TrawlSelect)
}
\description{
First sentence. second sentence.

\itemize{
  \item bla1. Dateframe of lon/lat coordinates corresponding to columns of sst$field (5 deg resolution)
  \item bla2. Vector of monthly date values corresponding to rows of sst$field
  \item bla3. Matrix of sea level pressure values by month and lon/lat position.
}
}
\examples{
data(data_TrawlSelect)
head(data_TrawlSelect)


}
\keyword{datasets}
\keyword{length}

