% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datThen.R
\docType{data}
\name{datThen}
\alias{datThen}
\title{Data of Then's study}
\format{
A data frame of class lfq consisting of:
\itemize{
  \item \strong{M} natural mortality,
  \item \strong{Linf} asymptotic length of VBGF,
  \item \strong{K} growth coefficient of VBGF
}
}
\source{
Then, A. Y., J. M. Hoenig, N. G. Hall, D. A. Hewitt. 2015. Evaluating the predictive
performance of empirical estimators of natural mortality rate using information on over
200 fish species. ICES J. Mar. Sci. 72: 82-92.
}
\usage{
data(datThen)
}
\description{
Original data of Then's study
}
