% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{emperor}
\alias{emperor}
\title{Emperor data}
\format{A dataframe containing: 1. \strong{year} years, 2. \strong{Y} yield,
   3. \strong{CPUE} CPUE or abundance index}
\source{
Dharmendra, D., Solmundsson, J., 2005. Stock assessment of the offshore Mauritian banks
using dynamic biomass models and analysis of length frequency of the Sky Emperor
(\emph{Lethrinus mahsena}). Fisheries Training Program The United Nations University, 61
}
\usage{
data(emperor)
}
\description{
This dataset contains information about sky emperor (\emph{Lethrinus mahsena})
 and its fisheries of offshore Mauritius banks (Nazareth banks). It can be used for
 production models (\code{\link{prod_mod}} and \code{\link{prod_mod_ts}}).
}
\examples{
data(emperor)
str(emperor)
summary(emperor)


}
\keyword{data}
\keyword{dataset}
\keyword{emperor}
