% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.ageBevertonHoltsZ.R
\docType{data}
\name{ex.ageBevertonHoltsZ}
\alias{ex.ageBevertonHoltsZ}
\title{Example dataset of length for use in BevertonHoltsZ}
\format{A list consisting of: 1. a dataframe for lon/lat positions, 2. a vector of
monthly date values, and 3. a matrix of sst anomaly values by month and lon/lat position
(1906 rows, 264 columns)}
\usage{
data(ex.ageBevertonHoltsZ)
}
\description{
First sentence. second sentence.

\itemize{
  \item bla1. Dateframe of lon/lat coordinates corresponding to columns of sst$field (5 deg resolution)
  \item bla2. Vector of monthly date values corresponding to rows of sst$field
  \item bla3. Matrix of sea level pressure values by month and lon/lat position.
}
}
\examples{

### Ex 1. Plot of single month
data(ex.ageBevertonHoltsZ)
head(ex.ageBevertonHoltsZ)


}
\keyword{datasets}
\keyword{length}

