% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.CatchCurve.R
\docType{data}
\name{ex.CatchCurve}
\alias{ex.CatchCurve}
\title{Example dataset of length for use in CatchCurve}
\format{A list consisting of: 1. a dataframe for lon/lat positions, 2. a vector of
monthly date values, and 3. a matrix of sst anomaly values by month and lon/lat position
(1906 rows, 264 columns)}
\source{
\url{http://www.esrl.noaa.gov/psd/data/gridded/data.kaplan_sst.html}
}
\usage{
data(ex.CatchCurve)
}
\description{
First sentence. second sentence.

\itemize{
  \item bla1. Dateframe of lon/lat coordinates corresponding to columns of sst$field (5 deg resolution)
  \item bla2. Vector of monthly date values corresponding to rows of sst$field
  \item bla3. Matrix of sea level pressure values by month and lon/lat position.
}
}
\examples{

### Ex 1. Plot of single month
data(ex.CatchCurve)
head(ex.CatchCurve)


}
\keyword{datasets}
\keyword{length}

