% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{ex.LengthCC}
\alias{ex.LengthCC}
\title{Exemplary length composition data}
\format{A dataframe consisting of: 1.midLengths, and 2.catch (18 rows, 2 columns)}
\source{
\url{Ziegler or Sparre?}
}
\usage{
data(ex.LengthCC)
}
\description{
Exemplary length composition data for Upeneus vittatus from Manila Bay, Philippines (from Ziegeler,1979) for the application of the length converted catch curve. Corresponding Linf and K are, 23.1cm and 0.59 per year, respectively. The \code{ex.LengthCC} data set is bull

\itemize{
  \item midLengths Midpoints of length classes
  \item catch Catch in Numbers
}
}
\examples{
### Ex 1.
data(ex.LengthCC)
output <- LengthConCatchCurve(midLengths = ex.LengthCC[,1], catch = ex.LengthCC[,2], Linf = 23.1, K = 0.59)
 output
}
\keyword{datasets}

