% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ex.Z_CPUE.R
\docType{data}
\name{ex.Z_CPUE}
\alias{ex.Z_CPUE}
\title{Example dataset of length for the calculation of Z form CPUE data}
\format{A list consisting of: 1. a dataframe for lon/lat positions, 2. a vector of
monthly date values, and 3. a matrix of sst anomaly values by month and lon/lat position
(1906 rows, 264 columns)}
\source{
\url{http://www.esrl.noaa.gov/psd/data/gridded/data.kaplan_sst.html}
}
\usage{
data(ex.Z_CPUE)
}
\description{
from exercise 4.3

\itemize{
  \item bla1. Dateframe of lon/lat coordinates corresponding to columns of sst$field (5 deg resolution)
  \item bla2. Vector of monthly date values corresponding to rows of sst$field
  \item bla3. Matrix of sea level pressure values by month and lon/lat position.
}
}
\examples{
### Ex 1.
data(ex.Z_CPUE)
head(ex.Z_CPUE)
}
\keyword{datasets}
\keyword{length}

