% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillnetfit.R
\name{gillnetfit}
\alias{gillnetfit}
\title{Millar's original gillnet selectivity fitting function}
\source{
https://www.stat.auckland.ac.nz/~millar/selectware/
}
\usage{
gillnetfit(data, meshsizes, type = "norm.loc", rel = NULL,
  plots = c(FALSE, FALSE), plotlens = NULL, details = FALSE)
}
\arguments{
\item{data}{a matrix}

\item{meshsizes}{meshsizes}

\item{type}{type of dist}

\item{rel}{relative power}

\item{plots}{include plots}

\item{plotlens}{plot lengths}

\item{details}{include details in output}
}
\value{
list of fitted arameters
}
\description{
Millar's original gillnet selectivity fitting function
}
\references{
Millar, R. B., Holst, R., 1997. Estimation of gillnet and hook selectivity
 using log-linear models. \emph{ICES Journal of Marine Science: Journal du Conseil},
 54(3):471-477
}

