% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GillnetSelect.R, R/GillnetSelect2.R
\name{GillnetSelect}
\alias{GillnetSelect}
\title{Gillnet selctivity}
\usage{
GillnetSelect(param)

GillnetSelect(param)
}
\arguments{
\item{param}{A list with following parameters: midlengths of size classes (\code{$midLengths}), number of fish caught with net 1 \code{$numNet1}, number of fish caught with net 2 (\code{$numNet2}), and the meshsizes of both nets (\code{$msNet1} & \code{$msNet2}).}

\item{classes}{Midpoints of the length class as vector (length frequency data) or ages as vector (age composition data).}

\item{catch}{Catch as vector, or a matrix with catches of subsequent years if the catch curve with constat time intervals should be applied.}

\item{numNet1}{Number of individuals caught using the smaller meshed net}

\item{numNet2}{Number of individuals caught using the larger meshed net}

\item{msNet1}{mesh size of smaller meshed net}

\item{msNet2}{mesh size of larger meshed net}
}
\description{
This function estimates the selecitvity of gillnets nets.

This function estimates the selecitvity of a gillnet from an experimental catch with two gillnets with different mesh sizes.
}
\details{
Preconditions: selection curves of two mesh sizes must overlap, nets set in same area, during the same time. To calculate selection factor (SF), L25, L50 and L75 for gillnets. Assumptions: optimum length Lm is proportional to mesh size (Lm = SF * m), two selection curves have the same standard deviation, nets have the same fishing power (same dimensions and material). Assumption that selection curves are normally distributed with a common standard deviation.

Preconditions: selection curves of two mesh sizes must overlap, nets set in same area, during the same time. To calculate selection factor (SF), L25, L50 and L75 for gillnets. Assumptions: optimum length Lm is proportional to mesh size (Lm = SF * m), two selection curves have the same standard deviation, nets have the same fishing power (same dimensions and material). Assumption that selection curves are normally distributed with a common standard deviation.
}
\examples{
\donttest{
data("ex.GillnetSelect")
with(ex.GillnetSelect,GillnetSelect(midLengths,numCa1,numCa2,
   msNet1 = 8.1,msNet2 = 9.1))
}

data("data_GillnetSelect")
output <- GillnetSelect(data_GillnetSelect)
plot(output)

}
\references{
XX

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

