% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{gotcha}
\alias{gotcha}
\title{'GOTCHA' method for aggregating catch numbers by pseudocohort}
\usage{
gotcha(
  lfq,
  n.pseudocohort = dim(lfq$catch)[1],
  n.pseudocohort.per.yr = NULL,
  slice.reso = 52
)
}
\arguments{
\item{lfq}{an object of class lfq}

\item{n.pseudocohort}{numeric. Number of pseudocohorts (i.e. total slices).
Defaults to the number of length bins in the lfq object. Overridden
when \code{n.pseudocohort.per.yr} is defined.}

\item{n.pseudocohort.per.yr}{numeric. Number of pseudocohorts per year. For
longer-lived species, setting \code{n.pseudocohort.per.yr = 1} would
result in a single aggregate value per year class. When specified,
  \code{n.pseudocohort} and  \code{slice.reso} are overridden.}

\item{slice.reso}{numeric. When \code{n.pseudocohort} is defined (default),
\code{slice.reso} defines a fine level of cohort slices per year to be
used as an initial determination of birthdates before pseudocohort
aggregation. Overridden when \code{n.pseudocohort.per.yr} is defined.}
}
\value{
a data.frame with catch numbers (n) aggregated by pseudocohort
  (bday, rel.age).
}
\description{
The method is useful as a pre-cursor to a catch-curve analysis
  whereby total mortality can be estimated based on the slope of catch numbers
  (log-transformed) as a function of (relative) age. As demonstrated by Pauly
  (1990), the method can be applied to cases where seasonally-oscillating
  growth creates inconsistencies of age at length depending on the time of
  year. The method aggregates by pseudocohort (n), as opposed to by length class
  in the length-converted-catch-curve (LCCC), and therefore does not require
  the correction of dividing numbers by time spent in a given length bin
  (n/dt).
}
\examples{
# load and visualize fit and cohorts
lfq <- synLFQ4
lfq$par <- list(Linf = 80, K = 0.5, ta = 0.25, C = 0.75, ts = 0.5)
lfq <- lfqModify(lfq, bin_size = 2, years = 2006:2007)
lfq <- lfqRestructure(lfq, MA = 5)
plot(lfq, image.col = NA)
lfq <- lfqCohort(lfq, n.per.yr = 1, calc_dt = FALSE)
image(lfq$dates, lfq$midLengths, t(lfq$cohort),
  col = adjustcolor(rainbow(100),0.4), add = TRUE)

# default settings - n.pseudocohorts = number of length bins
res <- gotcha(lfq, slice.reso = 36)
plot(log(n) ~ rel.age, res)
ressub <- subset(res, subset = rel.age > 2)
points(log(n) ~ rel.age, ressub, pch = 16)
fit <- lm(log(n) ~ rel.age, data = ressub)
abline(fit)
legend("topright", legend = paste("Z =", round(-coef(fit)[2],2)), bty = "n")

# n.pseudocohorts = defined number of bins
res <- gotcha(lfq, n.pseudocohort = 16)
plot(log(n) ~ rel.age, res)
ressub <- subset(res, subset = rel.age > 2)
points(log(n) ~ rel.age, ressub, pch = 16)
fit <- lm(log(n) ~ rel.age, data = ressub)
abline(fit)
legend("topright", legend = paste("Z =", round(-coef(fit)[2],2)), bty = "n")

# n.pseudocohorts = one per year
res <- gotcha(lfq, n.pseudocohort.per.yr = 1)
plot(log(n) ~ rel.age, res)
ressub <- subset(res, subset = rel.age > 2)
points(log(n) ~ rel.age, ressub, pch = 16)
fit <- lm(log(n) ~ rel.age, data = ressub)
abline(fit)
legend("topright", legend = paste("Z =", round(-coef(fit)[2],2)), bty = "n")

}
\references{
Pauly, D. (1990). Length-converted catch curves and the seasonal growth
  of fishes. Fishbyte, 8(3), 33–38.
}
