% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_soVB.R
\name{growth_soVB}
\alias{growth_soVB}
\title{Seasonalised version of the von Bertalanffy growth equation}
\usage{
growth_soVB(Linf, K, t, t0, ts, C)
}
\arguments{
\item{Linf}{infinite length for investigated species in cm [cm]}

\item{K}{growth coefficent for investigated species per year [1/year]}

\item{t}{time}

\item{t0}{theoretical time zero, at which individuals of this species hatch}

\item{ts}{summer point}

\item{C}{strength of oscillation}
}
\value{
A vector with lengths
}
\description{
This is the seasonalised version of the von Bertalanffy growth equation
}
\details{

}
\examples{
t <- seq(0,6,0.1)
L <- growth_soVB(Linf=80, K=0.6, t=t, t0=-0.1, ts=0.5, C=0.75)
plot(t, L, t="l")


}
\references{
Somers, I. F. (1988). On a seasonally oscillating growth function. Fishbyte, 6(1), 8-11
}
\keyword{function}
\keyword{growth}

