% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_VB.R
\name{growth_VB}
\alias{growth_VB}
\title{Von Bertalanffy growth equation}
\usage{
growth_VB(Linf, K, t, t0 = 0, L0 = NA)
}
\arguments{
\item{Linf}{infinite length for investigated species in cm [cm]}

\item{K}{growth coefficent for investigated species per year [1/year]}

\item{t}{time}

\item{t0}{theoretical time zero, at which individuals of this species hatch}
}
\value{
A vector with lengths
}
\description{
This is the von Bertalanffy growth equation with t0 or L0 respectively
}
\details{

}
\examples{
# with t0
t <- seq(0,6,0.1)
L <- growth_VB(Linf=80, K=0.6, t=t, t0=-0.1)
plot(t, L, t="l")

# with L0
t <- seq(0,6,0.1)
L <- growth_VB(Linf=80, K=0.6, t=t, L0=2)
plot(t, L, t="l")


}
\references{

}
\keyword{function}
\keyword{growth}

