% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{Jack-knife method for confidence intervals for growth parameters}
\usage{
jackknife(elefanFit)
}
\arguments{
\item{elefanFit}{lfq list fitted with growth parameters estimated with ELEFAN_SA or ELEFAN_GA}
}
\description{
The jackknife routine can be used to estimate confidence intervals for
  the growth parameters estimated with ELEFAN.
}
\keyword{CI}
\keyword{function}
\keyword{jackknife}
