% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JonesVanZalingeMethod.R
\name{JonesVanZalingeMethod}
\alias{JonesVanZalingeMethod}
\title{Length converted cumulative catch curve (The Jones and van Zalinge method)}
\usage{
JonesVanZalingeMethod(midLengths, cumCatch, Linf, K, t0 = 0,
  catchCorFac = NA)
}
\arguments{
\item{midLengths}{Midpoints of the size class intervals as vector}

\item{cumCatch}{Cumulative catch per sampling time as matrix or the total cumulative catch as vector.}

\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficient for investigated species per year [1/year].}

\item{t0}{Theoretical time zero, at which individuals of this species hatch.}

\item{catchCorFac}{optional: Correction factor for catch, in case provided catch does spatially or temporarily not reflect catch for fishing ground of a whole year.}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted cumulative catch curve.
}
\details{
The length converted linearised cumulative catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
\donttest{
 data("ex.LengthCC")
 #calculating the cumulated catch of fish of length L and above:
 cumulativeCatch <- rev(cumsum(rev(ex.LengthCC[,2])))
 output <- JonesVanZalingeMethod(midLengths = ex.LengthCC[,1],
   cumCatch = cumulativeCatch, Linf = 23.1, K = 0.59)
 output
 }
}

