% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LengthConCatchCurve.R
\name{LengthConCatchCurve}
\alias{LengthConCatchCurve}
\title{Length converted catch curve}
\usage{
LengthConCatchCurve(midLengths, catch, Linf, K, t0 = 0, catchCorFac = NA)
}
\arguments{
\item{midLengths}{Midpoints of the length class as vector}

\item{catch}{Catch per sampling time as matrix or the total catch as vector.}

\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficent for investigated species per year [1/year].}

\item{t0}{Theoretical time zero, at which individuals of this species hatch.}

\item{catchCorFac}{optional: Correction factor for catch, in case provided catch does spatially or temporarily not reflect catch for fishing ground of a whole year.}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve.
}
\details{
The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
\donttest{
 data("ex.LengthCC")
 output <- LengthConCatchCurve(midLengths = ex.LengthCC[,1], catch = ex.LengthCC[,2], Linf = 23.1, K = 0.59)
 output
 }
}
\references{
Sparre ???
}

