% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqCohort.R
\name{lfqCohort}
\alias{lfqCohort}
\title{Assign cohort and relative age to bins}
\usage{
lfqCohort(lfq)
}
\arguments{
\item{lfq}{an lfq object with fitted VBGF parameters in `lfq$par` slot.}
}
\value{
a list of class `lfq` with additional list elements "cohort" and "rel.age"
}
\description{
Assign cohort and relative age to bins
}
\examples{
data(synLFQ4)
synLFQ4 <- lfqRestructure(synLFQ4, MA=11)
synLFQ4$par <- list(Linf = 80, K = 0.5, t_anchor = 0.25, C = 0.75, ts = 0.5)

plot(synLFQ4, Fname = "rcounts", ylim = c(0, max(synLFQ4$midLengths)))
PAR2 <- synLFQ4$par
PAR2$t_anchor <- (PAR2$t_anchor - 0.5) \%\% 1
tmp <- lfqFitCurves(synLFQ4, par = PAR2,
  draw = TRUE, col=8, lty=1
)
synLFQ4 <- lfqCohort(synLFQ4)
pal <- colorRampPalette(c(4,5,7,2))
with(synLFQ4, image(x = dates, y = midLengths, z = t(cohort),
  col = adjustcolor(pal(max(cohort, na.rm = TRUE)), 0.5),
  add=TRUE
))

### Use all data in catch curve
df <- data.frame(
  rel.age = c(synLFQ4$rel.age[,-25]),
  n = c(synLFQ4$catch[,-25]),
  cohort = c(synLFQ4$cohort[,-25])
)
df$age <- floor(df$rel.age)

# floor of rel.age
agg <- aggregate(n ~ age, data = df, FUN = sum)
plot(log(n) ~ age, agg)
fit <- lm(log(n) ~ age, subset(agg, age \%in\% seq(1,4)))
abline(fit)
coef(fit)[2]

# direct rel.age
agg <- aggregate(n ~ rel.age, data = df, FUN = sum)
plot(log(n) ~ rel.age, agg)
fit <- lm(log(n) ~ rel.age, subset(agg, rel.age > 0.75 & rel.age < 3))
abline(fit)
-coef(fit)[2]


### For a single cohort
dfsub <- subset(df, cohort == 6)# most complete cohort

# direct rel.age
agg <- aggregate(n ~ rel.age, data = dfsub, FUN = sum)
plot(log(n) ~ rel.age, agg)
fit <- lm(log(n) ~ rel.age, subset(agg, rel.age > 1 & rel.age < 3))
abline(fit)
-coef(fit)[2]

}
