% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFANboot.R
\name{lfqResample}
\alias{lfqResample}
\title{Resampling of length-frequency data}
\usage{
lfqResample(lfq, boot = NULL)
}
\arguments{
\item{lfq}{a length frequency object of the class `lfq`.}

\item{boot}{optional; an object of class 'lfqBoot' for reproducing
a lfq data by re-using seed values}
}
\value{
a resampled version of the `lfq` class dataset.
}
\description{
The function resamples the `lfq` data by sample date.
  Sampling is done in a non-parametric manner that follows the relative
  frequencies of the original data, allowing for individual counts to be
  selected more than once (i.e. `replace = TRUE` in \link[base]{sample}),
  and resulting in total counts (by sample) equal to the original data.
}
\examples{
# load data
data(alba)

# permutate data
alba_p <- lfqResample(alba)

# side-by-side plot
op <- par( mfcol = c(2,1), mar=c(4,4,2,1) )
plot(lfqRestructure(alba)); mtext("original", side=3, line=0.25)
plot(lfqRestructure(alba_p)); mtext("resampled", side=3, line=0.25)
box()
par(op)

# relative difference
alba_diff <- alba
alba_diff$delta <- (alba$catch - alba_p$catch) / alba$catch
alba_diff$delta[is.na(alba_diff$delta)] <- 0
with(alba_diff, image(x=dates, y=midLengths, z=t(delta), col=rev(cm.colors(21))) )
with(alba_diff, contour(x=dates, y=midLengths, z=t(delta), add=TRUE))

}
