% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqRestructure_logspline.R
\name{lfqRestructure_logspline}
\alias{lfqRestructure_logspline}
\title{Restructuring of length frequency data using log-transformed frequencies}
\usage{
lfqRestructure_logspline(param, n.mult = 5, n.peak = 5,
  endrule = "median", logscore = FALSE)
}
\arguments{
\item{param}{a list consisting of following parameters:
\itemize{
  \item \strong{midLengths} midpoints of the length classes
  \item \strong{dates} dates of sampling times (class Date)
  \item \strong{catch} matrix with catches/counts per length class (row) and
  sampling date (column)
}}

\item{n.mult}{multiplier to use in expanding the number of bins (Default=5)}

\item{n.peak}{expected number of cohorts (Default=5)}

\item{endrule}{running median rule (Default="median")}

\item{logscore}{Logical. Return log-transformed scores (Default=FALSE)}
}
\description{
This is an alternate (experimantal) LFQ restructuring method
that works on log-transformed frequencies, followed by a calculation
of deviations againsta a fitted spline function.
}
\examples{
data("synLFQ4")

synLFQ4 <- lfqRestructure_logspline(synLFQ4, n.peak=5)
plot(synLFQ4)

synLFQ4 <- lfqRestructure_logspline(synLFQ4, n.peak=7, logscore=TRUE)
plot(synLFQ4)

# compare to lfqRestructure
op <- par(mfcol=c(2,1), mar=c(4,4,1,1))
synLFQ4 <- lfqRestructure(synLFQ4, MA=13)
plot(synLFQ4)
synLFQ4 <- lfqRestructure_logspline(synLFQ4, n.peak=7, logscore=FALSE)
plot(synLFQ4)
par(op)

}
