% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MillarsGillnetSelect.R
\name{MillarsGillnetSelect}
\alias{MillarsGillnetSelect}
\title{Millar's gillnet selctivity}
\usage{
MillarsGillnetSelect(param, model = "normal_fixed", rel = NULL,
  plotlens = NULL)
}
\arguments{
\item{param}{A list with following parameters: vector with midlengths of size classes
(\code{$midLengths}), vector with meshSizes in increasing order (\code{$meshSizes}),
and a matrix with the number of individuals caught with each sized mesh
(\code{$CatchPerNet_mat}).}

\item{model}{A character string indicating which type fo model for the estimation
of the selection curves should be used: \code{"fixed_normal"}, \code{"normal"},
\code{"gamma"}, or \code{"lognormal"}.}

\item{rel}{Vector of realtive efficencies}

\item{plotlens}{Vector of lengths for which values of relative retention are required}
}
\description{
This function estimates the selecitvity of a gillnet from an experimental
   catch with two gillnets with different mesh sizes.
}
\details{
This function estimates the fractions retained by each net (SNet1 & SNet2), the
  optimum lengths for each net, the selection factor (SF), and the standard deviation
  of the factor (stand.dev).
  Assumptions of this method are, that (i) the optimum length Lm is proportional to the mesh
  size (Lm = SF * m), (ii) the selection curves are normally distributed with a common
  standard deviation, (iii) the nets have the same fishing power (same dimensions and material).
  Requirements for the experimental set-up are: selection curves corresponding to the two
  mesh sizes have to overlap, and the nets have to be set in the same area, during the
  same time.
}
\examples{
# load data
data(data_MillarsGillnet)

# run model
output <- MillarsGillnetSelect(data_MillarsGillnet, model = "normal_fixed",
   plotlens = NULL, rel = NULL)

# investigate results
output

# plot results
plot(output)

}
\references{
Millar, R. B., Holst, R., 1997. Estimation of gillnet and hook selectivity
 using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.

 Holt, S. J. 1963. A method for determining gear selectivity and its application.
 ICNAF Special Publication, 5: 106–115.
}

