% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_empirical.R
\name{MstochThen}
\alias{MstochThen}
\title{Empirical M formula by Then in bootstrap framework}
\usage{
MstochThen(boot, CI = 95)
}
\arguments{
\item{boot}{an object of class 'lfqBoot'}

\item{CI}{percentage for confidence intervals (Default: 95)}
}
\description{
Resamples from Then's original data and refits
    original model within bootstrap framework to estimate M with confidence intervals.
}
\references{
Then, A. Y., J. M. Hoenig, N. G. Hall, D. A. Hewitt. 2015. Evaluating the predictive
performance of empirical estimators of natural mortality rate using information on over
200 fish species. ICES J. Mar. Sci. 72: 82-92.
}
