% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFANboot.R
\name{MultiMaxDen}
\alias{MultiMaxDen}
\title{Multivariate maximum density estimates from a bootstrapped object}
\usage{
MultiMaxDen(boot, CI = 95, omega = 1.5)
}
\arguments{
\item{boot}{object of class "lfqBoot"}

\item{CI}{confidence interval default 95%}

\item{omega}{multiplication factor defining the number of eval.points}
}
\value{
an object of class 'lfqBoot' with following objects
   \itemize{
   \item \strong{bootRaw} a data.frame of fitted VBGF parameters (columns) by
              permutation (rows),
   \item \strong{seed} a vector with seed values used in the LFQ resampling process
   \item \strong{maxDen} a dataframe with the modes of the VBGF parameters,
   \item \strong{CI} a dataframe with the lower and upper 95% (default) confidence
              interval for the VBGF growth parameters.
   }
}
\description{
estimates maximum density estimates and confidence intervals for
  (seasonalised) VBGF parameters obtained with ELEFAN_GA_boot or ELEFAN_boot_SA
}
