% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PaulysM.R
\name{PaulysM}
\alias{PaulysM}
\title{Pauly's empirical formula for natural mortality (M)}
\usage{
PaulysM(Linf, K, Temp, schooling = FALSE)
}
\arguments{
\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficient for investigated species per year [1/year].}

\item{Temp}{Average annual temperature at the surface in degrees centigrade.}

\item{schooling}{Logical parameter accounting for schooling behaviour, according to Pauly (1983). Default is FALSE.}
}
\description{
Pauly's empirical formula to calculate natural mortality (M)
}
\details{
Pauly's empirical formula to calculate natural mortality (M). If accounting for schooling behaviour M is multiplied by 0.8 according to Pauly (1983).
}
\examples{
PaulysM(Linf = 80, K = 0.5, Temp = 25)
}

