% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.catchCurve.R
\name{plot.catchCurve}
\alias{plot.catchCurve}
\title{Plotting catch curve}
\usage{
\method{plot}{catchCurve}(x, plot_selec = FALSE, ...)
}
\arguments{
\item{x}{A list of the class \code{"catchCurve"} containing the results of the
catchCurve model.}

\item{plot_selec}{logical; if TRUE the regression line is plotted for not fully
exploited length groups and the probability of capture is plotted}

\item{...}{normal parameters from plot function}
}
\description{
This function plots the results from the \code{\link{catchCurve}} model.
}
\details{
A function to plot the results of the catchCurve model.
}
\examples{
\donttest{
data(synLFQ3)
output <- catchCurve(synLFQ3, calc_ogive = TRUE)
plot(output, plot_selec = TRUE)
}

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

