% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ELEFAN.R
\name{plot.ELEFAN}
\alias{plot.ELEFAN}
\title{ELEFAN plot}
\usage{
\method{plot}{ELEFAN}(x, ...)
}
\arguments{
\item{x}{A list of the class \code{"ELEFAN"} containing the results of ELEFAN}

\item{...}{Default parameter options from plot function}
}
\description{
This function plots the growth curves calculated by means of ELEFAN
   (\code{\link{ELEFAN}}).
}
\details{
This function draws a selectivity plot for the object class
   \code{"ELEFAN"}, which is created by applying ELEFAN
   \code{\link{ELEFAN}}.
}
\examples{
\donttest{
#load data
  Linf.p = 45.2
  K.p = 0.6
  tshift.p = 360/365

plot(output, Linf= , K=, )  # with entering tshift? or it looks up tshift from time matrix but what happens if user choose L and K values outside of time mat???

}

}
\references{
Pauly, D. and N. David, 1981. ELEFAN I, a BASIC program for the objective extraction of
growth parameters from length-frequency data. \emph{Meeresforschung}, 28(4):205-211

Pauly, D., 1987. A review of the ELEFAN system for analysis of length-frequency data in
fish and aquatic invertebrates. \emph{ICLARM Conf. Proc.}, (13):7-34

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

