% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MillarsGillnet.R
\name{plot.MillarsGillnet}
\alias{plot.MillarsGillnet}
\title{Millar's gillnet selectivity plot}
\usage{
\method{plot}{MillarsGillnet}(estim, ret_dev = c(T, T), cex = 1,
  title = "Deviance residuals", ...)
}
\arguments{
\item{estim}{A list of the class "GillnetSelect" containing the results of the gillnet selectivity function.}
}
\description{
This function plots the selectivity estimates of the function \code{\link{GillnetSelect}}.
}
\details{
A function to plot the results of the gillnet selectivity estimation.
}
\examples{
data(data_GillnetSelect3)
output <- MillarsGillnetSelect(data_GillnetSelect3, model = "normal_fixed",
   plotlens = NULL, rel = NULL)
output
plot(output)

}
\references{
Millar, R. B., Holst, R., 1997. Estimation of gillnet and hook selectivity
 using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.
}

