% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predict_mod.R
\name{plot.predict_mod}
\alias{plot.predict_mod}
\title{Plotting prediction models yield per recruit and Thompson & Bell}
\usage{
\method{plot}{predict_mod}(x, type = "ypr", xaxis1 = "FM", yaxis1 = "Y_R",
  yaxis2 = "B_R", ...)
}
\arguments{
\item{x}{a object of the class 'predict_mod',}

\item{...}{optional parameters of plot function}
}
\description{
This function plots objects of the class "predict_mod", which are results
   of the function \code{\link{predict_mod}}.
}
\examples{


}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

