% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.prod_mod_dyn.R
\name{plot.prod_mod_dyn}
\alias{plot.prod_mod_dyn}
\title{Plotting dynamic production models}
\usage{
\method{plot}{prod_mod_dyn}(x, ...)
}
\arguments{
\item{x}{a object of the class 'prod_mod_dyn',}

\item{...}{optional parameters of plot function}
}
\description{
This function plots objects of the class "prod_mod_dyn".
}
\examples{
# load data
data(emperor)

# run model
output <-  prod_mod_dyn(data = emperor)

# plot output
plot(output)


}
\references{
Dharmendra, D., Sólmundsson, J., 2005. Stock assessment of the offshore Mauritian banks
using dynamic biomass models and analysis of length frequency of the Sky Emperor
(\emph{Lethrinus mahsena}). Fisheries Training Program The United Nations University, 61
}

