% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFANboot.R
\name{plotBoot}
\alias{plotBoot}
\title{Bootstrapped VBGF estimates for the alba lfq dataset}
\usage{
plotBoot(lfq, boot, pType = "lfq", display = "CI", best = "maxDen",
  Fname = "catch", pCCori = FALSE, pRef = "fmax", par = NULL,
  agemax = NULL, rel = FALSE, curve.col = 1, hist.sc = 0.5,
  hist.col = c("white", "black", "orange", "darkgreen"), image.col = NULL,
  region.col = NULL, zlim = NULL, zlimtype = "balanced",
  date.axis = "traditional", date.at = seq(as.Date("1500-01-01"),
  as.Date("2500-01-01"), by = "months"), date.format = "'\%y-\%b",
  xlab = "default", ylab = "default", draw = TRUE, ylim = NULL,
  xlim = NULL, CI = 95, add_legend = TRUE, add_max_dens_legend = TRUE,
  points.col = "black", points.pch = 19, perm.col = adjustcolor("grey50",
  0.1), perm.lwd = 1, ci.col = 1, ci.lty = 2, ci.lwd = 1,
  maxd.col = 1, maxd.lty = 1, maxd.lwd = 2, ...)
}
\description{
Bootstrapped VBGF estimates for the alba lfq dataset as estimated by
  \code{\link[TropFishR]{ELEFAN_GA}}.
}
\examples{
\donttest{
#### For documentation only - How data was produced ####
library(parallel)
library(TropFishR)
library(ks)

data(alba)

# ELEFAN_GA_boot settings
MA <- 7
low_par <- list(Linf = 8, K = 0.1, t_anchor = 0, C = 0, ts = 0)
up_par <- list(Linf = 15, K = 5, t_anchor = 1, C = 1, ts = 1)
seasonalised <- FALSE
popSize <- 100
maxiter <- 50
run <- 10
pmutation <- 0.2
nresamp <- 200

# Bootstrapped estimates
alba_boot <- ELEFAN_GA_boot(lfq=alba, MA = MA, seasonalised = seasonalised,
  up_par = up_par, low_par = low_par,
  popSize = popSize, maxiter = maxiter, run = run, pmutation = pmutation,
  nresamp = nresamp, parallel = TRUE, no_cores = detectCores(),
  seed = 20
)
}

plotBoot(lfq=alba, boot=alba_boot, plot="catchCurve", display="CI")
plotBoot(lfq=alba, boot=alba_boot, plot="catchCurve", display="bootRaw")

}
