% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mod.R
\name{predict_mod}
\alias{predict_mod}
\title{Prediction models: YPR and Thompson & Bell model}
\usage{
predict_mod(param, FM_change = NA, Lc_tc_change = NULL, type,
  stock_size_1 = NA, unit.time = "year", plus.group = NA, Linf, K,
  t0 = 0, s_list)
}
\arguments{
\item{param}{a list consisting of following parameters:
\itemize{
  \item \strong{age} or \strong{midLengths}: midpoints of the length class as vector (length-frequency
  data) or ages as vector (age composition data),
  \item \strong{Linf}: infinite length for investigated species in cm [cm],
  \item \strong{K}: growth coefficent for investigated species per year [1/year],
  \item \strong{t0}: theoretical time zero, at which individuals of this species hatch,
  \item \strong{catch}: catch as vector, or a matrix with catches of subsequent years if
  the catch curve with constat time intervals should be applied;
}}

\item{FM_change}{vector with ascending Fishing mortalities}

\item{Lc_tc_change}{vector with ascending Lc values}

\item{type}{indicating which model should applied: \code{"ypr"} for Beverton and Holt's
yield per recruit model and \code{"ThompBell"} for the Thompson and Bell model}
}
\value{
A list with the input parameters and dependent on the model type following
   list objects:
\itemize{
  \item \code{type = 'ypr'}
  \itemize{
     \item \strong{FM}: fishing mortalities,
     \item \strong{Lc} or \strong{tc}: lengths or ages at first capture,
     \item \strong{list_Lc_runs} or \strong{list_tc_runs}: a list with the dataframes for each Lc or tc value:
     \itemize{
     \item FM_change
     \item Y_R
     \item Y_R.rel
     \item B_R
     \item B_R.percent
     \item Ty
     \item Wy
     }
  }
  \item \code{type = 'ThomBell'}
  \itemize{
     \item \strong{dt}: delta t,
     \item \strong{N}: population number,
     \item \strong{dead}: deaths due to natural reasons,
     \item \strong{C}: catch,
     \item \strong{Y}: yield,
     \item \strong{B}: biomass,
     \item \strong{V}: value,
     \item \strong{totals}: summed up values (total catch, total yield, total value, average biomass),
     \item \strong{tot.C}: total catches for different x factors,
     \item \strong{tot.Y}: total yield values for different x factors,
     \item \strong{tot.V}: total values for different x factors,
     \item \strong{meanB}: average biomasses for different x factors,
     \item \strong{Xfact}: fishing mortality changes;
  }
  \item \code{type = 'ThomBell'} and \code{Lc_tc_change} provided
  \itemize{
     \item \strong{FM_change}: fishing mortality changes,
     \item \strong{Lc_tc_change}: changes in length or age at first capture,
     \item \strong{Lt}: lengths at age,
     \item \strong{sel}: probability of capture,
     \item \strong{mat_FM_Lc_com.C}: catch matrix for all fishing mortality and Lc/tc combinations,
     \item \strong{mat_FM_Lc_com.Y}: yield matrix for all fishing mortality and Lc/tc combinations,
     \item \strong{mat_FM_Lc_com.V}: value matrix for all fishing mortality and Lc/tc combinations,
     \item \strong{mat_FM_Lc_com.B}: biomass matrix for all fishing mortality and Lc/tc combinations;
  }

}
}
\description{
This is a function predicting catches, yield, biomass and value for different
   fishing mortality scenarions (in combination with gear changes). It allows to choose if the
   calculations should be based upon Beverton and Holt's yield per recruit model or the
   more sophisticated Thompson and Bell model.
}
\details{
better to treat last group always as a plus group.....
   The Thompson and Bell model incorporates an iteration step simulating the stock by means
   of the function \code{\link{stock_sim}}. In the case changes in gear characteristics -
   here measured in terms of Lc or tc, the length or age at first capture, respectively -
   should be explored a list with selectivity information about the gear has provided and
   the prediction models make use of the selectivity function \code{\link{select_ogive}}.
}
\examples{
#______________________________________
#       Yiel Per Recruit (YPR)
#______________________________________
# age structured data
# Nemipterus marginatus
threadfin <- list(Winf = 286,K = 0.37, t0 = -0.2, M = 1.1, tr = 0.4)

# run model
predict_mod(threadfin, FM_change = seq(0,6,0.1),
   Lc_tc_change = seq(0.2,1,0.2), type = 'ypr')  #where it is maximal  = MSY

# Leiognathus spendens (Pauly 1980)
ponyfish <- list(Winf = 64, K = 1, t0 = -0.2, M = 1.8, tr = 0.2)

# run model
predict_mod(ponyfish, Lc_tc_change = c(0.2,0.3,1.0), type = 'ypr')

#______________________________________
# length structured data
# Xiphias gladius (Berkeley and Houde 1980)
swordfish <- list(Linf = 309, K = 0.0949, M = 0.18, a=0.0003, b=3, Lr = 90)  ## T_Lr , a, b ??? assumed

# run model
predict_mod(swordfish, Lc_tc_change = c(100,118,150,180), type = 'ypr')

####test: E <- seq(0,0.9,0.1) FM <- E * M / (1 - E)


#______________________________________
#      Thompson and Bell model
#______________________________________
# with age structured data
# load data
data(shrimps)

# run model
predict_mod(shrimps, FM_change = seq(0.1,3,0.1), type = 'ThompBell')

# create list with selectivity information
select.list <- list(selecType = 'knife_edge',  #or 'gillnet' or 'trawl_ogive'
   Lc = 34, tc = 5, selecDist = 'lognormal',    #or 'normal_fixed'
   mesh_size = 8.1, mesh_size1 = 9.1, select_p1 = 21.1, select_p2 = 23.8)

# run model
predict_mod(shrimps,FM_change = seq(0,3,0.2), Lc_tc_change = seq(24,44,2),
   type = 'ThompBell', unit.time = 'month', Linf = 50, K = 0.3, t0 = 0.01,
   s_list = select.list)

#______________________________________
# with length structured data
# load data
data(hake)

# run model
predict_mod(hake,FM_change = seq(0.1,3,0.1), type = 'ThompBell')


}
\references{
example 1 : Kuwait (Garcia and van Zalinge 1982)

Millar, R. B., & Holst, R. (1997). Estimation of gillnet and hook selectivity using
log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3):471-477
}
\keyword{function}
\keyword{prediction}

