% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_ThompsonBell.R
\name{Predict_ThompsonBell}
\alias{Predict_ThompsonBell}
\title{Thompson and Bell prediction model}
\usage{
Predict_ThompsonBell(classes, mean_weight, FM, Z, value = NA, Tr, datatype,
  stock_size_1 = NA, plus.group = NA, FM_change = NA, fleet_mat = NA,
  fleet_FM_change = NA, fleet_plot_name = NA)
}
\arguments{
\item{classes}{Midpoints of the length class as vector (length frequency data) or ages as vector (age composition data).}

\item{FM}{reference F-at-age-array}

\item{datatype}{Type of data which is used for analysis, either 'length' or 'age', for length frequency or age composition data, respectively}

\item{catch}{Catch as vector, or a matrix with catches of subsequent years if the catch curve with constat time intervals should be applied.}

\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficent for investigated species per year [1/year].}

\item{t0}{Theoretical time zero, at which individuals of this species hatch (default: 0).}

\item{Winf}{Infinite weight in grams!}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve or from age at length data with catch curve.
}
\details{
better to treat last group always as a plus group..... For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
data("ex.Predict_ThompsonBell")
with(ex.Predict_ThompsonBell,Predict_ThompsonBell(classes = age,
  mean_weight = meanWeight,value = valueGramm,FM = FM,Z = Z,
  datatype = 'age',Tr = 1,stock_size_1 = NA,plus.group = 12))

}
\references{
example 1 : Kuwait (Garcia and van zalinge 1982)
}

