% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_ThompsonBell1.R
\name{Predict_ThompsonBell1}
\alias{Predict_ThompsonBell1}
\title{Thompson and Bell prediction model}
\usage{
Predict_ThompsonBell1(param, unit.time = "year", stock_size_1 = NA,
  plus.group = NA)
}
\arguments{
\item{param}{A list containing all information}

\item{unit.time}{Indicates if the age groups are per month (\code{"month"}) or per year (\code{"year"}). Default: \code{"year"}}

\item{stock_size_1}{Stock size of smallest age/length group}

\item{plus.group}{Indicates age/length group, which should be turned into a plus group (i.e. all groups above are comprised in one group)}

\item{mean_weight}{Average weight of fish}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve or from age at length data with catch curve.
}
\details{
better to treat last group always as a plus group..... For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
#age-based
data("data_Predict_ThompsonBell")
#without plus group
Predict_ThompsonBell1(data_Predict_ThompsonBell)
#with plus group
Predict_ThompsonBell1(data_Predict_ThompsonBell,plus.group=11)

# length-based
data(hake)
output <- Predict_ThompsonBell1(param = hake, stock_size_1 = 98919.3)
output

}
\references{
example 1 : Kuwait (Garcia and van zalinge 1982)
  Millar, R. B., & Holst, R. (1997). Estimation of gillnet and hook selectivity using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.
}

