% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_ThompsonBell2.R
\name{Predict_ThompsonBell2}
\alias{Predict_ThompsonBell2}
\title{Thompson and Bell prediction model}
\usage{
Predict_ThompsonBell2(param, FM_change)
}
\arguments{
\item{param}{A list}

\item{FM_change}{Vector containing new FM values}

\item{mean_weight}{Average weight of fish}

\item{datatype}{Type of data which is used for analysis, either 'length' or 'age', for length frequency or age composition data, respectively}

\item{FM}{Fishing mortality}

\item{Z}{Total mortality}

\item{value}{Information about the value/price of fish per kilo or gramm}

\item{Tr}{Age of recruitment}

\item{stock_size_1}{Stock size to start with}

\item{plus.group}{Should a plus group be created, by default (NA) not. BUt authors advise to do so}

\item{fleet_mat}{Matrix providing catch or FM values separated by fleets}

\item{fleet_unit}{Either 'Catch' or 'FM' indicating in which unit the fleet_mat is provided}

\item{fleet_FM_change}{Matrix containing new FM values for each fishery}

\item{fleet_plot_name}{Name of fishery, which should be used for plotting Yield per recurit against changes of FM}

\item{FM}{reference F-at-age-array}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve or from age at length data with catch curve.
}
\details{
better to treat last group always as a plus group..... For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{

# with age data
data(data_Predict_ThompsonBell)
Predict_ThompsonBell2(data_Predict_ThompsonBell,seq(0.1,3,0.1))

# with length data
data(hake)
Predict_ThompsonBell2(hake,seq(0.1,3,0.1))

}
\references{
example 1 : Kuwait (Garcia and van zalinge 1982)
  Millar, R. B., & Holst, R. (1997). Estimation of gillnet and hook selectivity using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.
}

