% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_ThompsonBell_Fleets.R
\name{Predict_ThompsonBell_Fleets}
\alias{Predict_ThompsonBell_Fleets}
\title{Thompson and Bell prediction model}
\usage{
<<<<<<< HEAD:man/Predict_ThompsonBell.Rd
Predict_ThompsonBell(classes, mean_weight, FM, Z, value = NA, Tr, datatype,
  stock_size_1 = NA, plus.group = NA, FM_change = NA, fleet_mat = NA,
<<<<<<< HEAD
  fleet_unit = NA, fleet_FM_change = NA, fleet_plot_name = NA)
=======
  fleet_unit = NA, fleet_FM_change = NA, fleet_plot_name = NA,
  s_list = NA, Linf = NA, K = NA, t0 = 0, L50 = NA, L75 = NA)
>>>>>>> 1536cdd9b2668bf517ddbb56e372d9add9ffb9d2
=======
Predict_ThompsonBell_Fleets(param, FM_change, fleet_mat, fleet_unit,
  fleet_FM_change, fleet_plot_name = NA, unit.time = "year",
  stock_size_1 = NA, plus.group = NA)
>>>>>>> 5138cd9cc43af0588ed99bd0456e59c932fdadc0:man/Predict_ThompsonBell_Fleets.Rd
}
\arguments{
\item{param}{A list}

\item{FM_change}{Vector containing new FM values}

\item{fleet_mat}{Matrix providing catch or FM values separated by fleets}

\item{fleet_unit}{Either 'Catch' or 'FM' indicating in which unit the fleet_mat is provided}

\item{fleet_FM_change}{Matrix containing new FM values for each fishery}

\item{fleet_plot_name}{Name of fishery, which should be used for plotting Yield per recurit against changes of FM}

\item{stock_size_1}{Stock size to start with}

\item{plus.group}{Should a plus group be created, by default (NA) not. BUt authors advise to do so}

\item{FM}{reference F-at-age-array}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve or from age at length data with catch curve.
}
\details{
better to treat last group always as a plus group..... For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{

# load data
data(data_Predict_ThompsonBell)

# construct fleet matrix
fleet_mat=matrix(ncol = 2,nrow = 12)
fleet_mat[,1] <- c(0.72,0.96,0.84,0.48,0.6,0.48,1.08,0.48,0.084,0.12,0.24,0.24)
fleet_mat[,2] <- c(0.48,0.36,0.48,0.96,1.32,0.72,0.48,0.72,1.116,1.68,2.52,2.28)
colnames(fleet_mat) <- c("artisanal", "industrial")

# construct matrix with changes of fishing mortalities by fleet
fleet_FM_change = matrix(ncol = 2,nrow = 8)
fleet_FM_change[,1] <- rep(1,8)
fleet_FM_change[,2] <- c(0,0.4,0.8,1.0,1.2,1.5,2,3)
colnames(fleet_FM_change) <- c("artisanal", "industrial")

# run the model
output <- Predict_ThompsonBell_Fleets(data_Predict_ThompsonBell,seq(0,3,0.2),
   fleet_mat,"FM",fleet_FM_change,fleet_plot_name = "industrial",
   'month',plus.group = 12)

# investigate results
output

}
\references{
example 1 : Kuwait (Garcia and van zalinge 1982)
  Millar, R. B., & Holst, R. (1997). Estimation of gillnet and hook selectivity using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.
}

