% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_ThompsonBell_FM_Lc.R
\name{Predict_ThompsonBell_FM_Lc}
\alias{Predict_ThompsonBell_FM_Lc}
\title{Thompson and Bell prediction model}
\usage{
Predict_ThompsonBell_FM_Lc(param, FM_change, stock_size_1 = NA,
  unit.time = "year", plus.group = NA, Linf, K, t0 = 0, s_list, Lc_change)
}
\arguments{
\item{param}{A list}

\item{FM_change}{Vector containing new FM values}

\item{stock_size_1}{Stock size to start with}

\item{unit.time}{Time unit "year" or "month", Default "year"}

\item{plus.group}{Should a plus group be created, by default (NA) not. BUt authors advise to do so}

\item{Linf}{growth parameter 1}

\item{K}{growth parameter 2}

\item{t0}{growth parameter 3}

\item{s_list}{List wit all selectivity information}

\item{Lc_change}{vector with ascending Lc values}
}
\description{
This is a function to calculate the total mortality (Z) from
   length composition data via the length converted catch curve or from age at
   length data with catch curve.
}
\details{
better to treat last group always as a plus group..... For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{

# load data
data(data_Predict_ThompsonBell)

# create list with selectivity information
select.list <- list(selecType = 'knife_edge',  #or 'gillnet' or 'trawl_ogive'
   Lc = 34,tc = 5,selecDist = 'lognormal',    #or 'normal_fixed'
   mesh_size = 8.1,mesh_size1 = 9.1,select_p1 = 21.1,select_p2 = 23.8)

# run model
Predict_ThompsonBell_FM_Lc(data_Predict_ThompsonBell,FM_change = seq(0,3,0.2),
  unit.time = 'month', Linf=50, K=0.3, t0=0.01, s_list=select.list,
  Lc_change=seq(24,44,2))


}
\references{
example 1 : Kuwait (Garcia and van zalinge 1982)
  Millar, R. B., & Holst, R. (1997). Estimation of gillnet and hook selectivity using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.
}

