% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepLFQ.R
\name{prepLFQ}
\alias{prepLFQ}
\title{Prepare lfq data for further analysis}
\usage{
prepLFQ(lfq, par = NULL)
}
\arguments{
\item{lfq}{lfq object with dates, midLengths, and catch}

\item{par}{growth parameter as resulting from e.g. \code{\link{ELEFAN}}}
}
\value{
lfq object with rearranged catch matrix (yearly sums) and growth parameters
   if provided.
}
\description{
Rearrange catch matrix in length frequency data (lfq class) to
   have one column per year. This is required for e.g. \code{\link{catchCurve}}.
}
\examples{
data(synLFQ4)
newlfq <- prepLFQ(synLFQ4)

}
\keyword{function}
\keyword{length-frequency}
\keyword{lfq}

