% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_mod_dyn.R
\name{prod_mod_dyn}
\alias{prod_mod_dyn}
\title{Non-equilibrium dynamic biomass models}
\usage{
prod_mod_dyn(param)
}
\arguments{
\item{param}{a list of parameters}
}
\description{
Non-equilibrium dynamic biomass models
}
\details{
MSY
}
\examples{
\donttest{
  # Y= Yield(catch) in tons I= abundance index (cpue)kg
  #SAYA Catch and Effort data 1989-2004 DYNAMIC BIOMASS MODEL
  yrs <- 1989:2004
  Y <- c( 2177, 1410, 1782, 2825, 3173, 3142, 2957, 2283, 1798, 2054, 2107,
     2099, 1283, 2090, 2354, 1689)
  I <- c( 74.3, 73.0, 88.0, 67.7, 69.1, 66.6, 67.0, 57.8, 71.8, 75.9, 70.1,
     77.8, 124.1, 83.3, 80.1, 71.2)

  #NAZARETH catch and effort data DYNAMIC BIOMASS MODEL 1989-2004
  yrs <- 1989:2004
  Y <- c( 837,914,793,952,1358,1494,1533,1253,1720,1086,1121,1080,1366,918,468,855)
  I <- c( 75.7,78.5,81.3,78.2,66.1,66.5,64.3,52.5,66.1,81.4,76.2,90.5,99.2,93.3,
     72.8,84.2)
}

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.



#@export
}

