% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmtool.R
\name{r2over3}
\alias{r2over3}
\title{2 over 3 rule to estimate r of HCRs}
\usage{
r2over3(indmat)
}
\arguments{
\item{indmat}{matrix with indices of last 5 years}
}
\description{
Apply 2 over 3 rule to estimate r of HCRs
}
