% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurves.R
\name{rcurves}
\alias{rcurves}
\title{Predict gillnet selectivity (old Millar method)}
\usage{
rcurves(type, meshsizes, rel, pars, plotlens)
}
\arguments{
\item{type}{("norm.loc", "norm.sca", "gamma", "lognorm")}

\item{meshsizes}{mesh sizes}

\item{rel}{relative powers (one for each mesh size)}

\item{pars}{selection curve parameters}

\item{plotlens}{vector of new mesh sizes for selectivity prediction}
}
\value{
selectivities
}
\description{
Predict gillnet selectivity (old Millar method)
}

