% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurves_Millar.R
\name{rcurves_Millar}
\alias{rcurves_Millar}
\title{Predict gillnet selectivity (old Millar method)}
\source{
https://www.stat.auckland.ac.nz/~millar/selectware/
}
\usage{
rcurves_Millar(type, meshsizes, rel, pars, plotlens)
}
\arguments{
\item{type}{("norm.loc", "norm.sca", "gamma", "lognorm")}

\item{meshsizes}{mesh sizes}

\item{rel}{relative powers (one for each mesh size)}

\item{pars}{selection curve parameters}

\item{plotlens}{vector of new mesh sizes for selectivity prediction}
}
\value{
selectivities
}
\description{
Predict gillnet selectivity (old Millar method)
}
\references{
Millar, R. B., Holst, R., 1997. Estimation of gillnet and hook selectivity
 using log-linear models. \emph{ICES Journal of Marine Science: Journal du Conseil},
 54(3):471-477
}

