% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment}
\alias{recruitment}
\title{Recruitment patterns}
\usage{
recruitment(param, Linf, K, t0)
}
\arguments{
\item{param}{a list}
}
\description{
This function estimates recrutiment patterns from length-frequency data-
}
\details{
should be possible to extract recruitment information from length frequency data (Pauly 1982),
but two main assumptions seldomly met:
 1. all fish in a data set grow as described by a single set of growth parameters
2. one month out of twelve always has zero recruitment (not met, because in temperate waters probable that more than one month zero recruitment and in tropics maybe always recruitment)

Recuritment patterns obtained by backward projection onto the length axis of a set of
length frequency data. The steps are:
 1.
projection onto the time axis of the frequencies after they have been divided by delta t,
the time needed to grow through the length class. (Alternatively, FiSAT allows,
following Moreau and Cuende (1991) the use of restructured frequencies (as used for
untransformed L/F data are used, thus partly compensating for the "spread"
illustrated in Fig. 5.19.);

2.
summation for each month (and irrespective of year) of the adjusted frequencies
projected onto each month;
3.
subtraction, from each monthly sum, of the lowest monthly sum to obtain a zero
value where apparent recruitment is lowest;
4.
output of monthly relative recruitment as a percentage of annual recruitment.

results pertain to months of the year only if precise etimate for t0 is given. (time
lag between peak spawning and recruitment can be used ot obtain value for t0)
}
\examples{
data(trout)

trout
Linf = 16
K = 0.4

}
\references{
Moreau, J., & Cuende, F. X., 1991. On improving the resolution of the recruitment
patterns of fishes. \emph{Fishbyte}, 9(1), 45-46.

Pauly, D., 1982. Studying single-species dynamics in a tropical multispecies context.
In Theory and management of tropical fisheries. \emph{ICLARM Conference Proceedings}
(Vol. 9, No. 360, pp. 33-70).

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.

#@export
}

