% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment2.R
\name{recruitment2}
\alias{recruitment2}
\title{Calculate recruitment pattern of lfq object}
\usage{
recruitment2(lfq, Lrecr = 0, plot = TRUE, hide.progressbar = FALSE)
}
\arguments{
\item{lfq}{a length-frequency object (i.e. class \code{lfq})}

\item{Lrecr}{numeric. Length at recruitment
(default: \code{Lrecr = 0}).}

\item{plot}{logical. Should monthly recruitment pattern by plotted
(default: \code{plot = TRUE}).}

\item{hide.progressbar}{logical. Should progress bar be displayed
(default: \code{hide.progressbar = FALSE})}
}
\value{
object of "histogram" class containing
monthly recruitment pattern
}
\description{
\code{recruitment2} calculates the recruitment pattern (i.e. monthly)
of an lfq object \code{\link{VBGF}} parameters and the length at
recruitment (\code{Lrecr}, e.g. to the population or the fishery).
Time at recruitment \code{trecr} is calculated via a hindcast
using the \code{\link{calc_trecr}} function.
}
\details{
This function does not represent a robust statistical method as it
assumes the same growth parameters for all individual counts
in the lfq count bins.
Nevertheless, the results should provide general information about
the recruitment pattern, e.g. relative recruitment strength by month.
}
\examples{
lfq <- synLFQ4
lfq$par <- list(
  Linf = 80, K = 0.5, t_anchor = 0.25, C = 0.75, ts = 0.5
)
tmp <- recruitment2(lfq = lfq)
tmp

}
