% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchaeferFox.R
\name{SchaeferFox}
\alias{SchaeferFox}
\title{Schaefer and Fox models}
\usage{
SchaeferFox(Y, f)
}
\arguments{
\item{Y}{Yield as vector (catch in weight)}

\item{f}{Fishing effort as vector}
}
\description{
Schaefer and Fox models are examples of so called surplus production
   models, and are used to calculate MSY and BMSY and so on.
   First sentence. second sentence.
}
\details{
MSY
}
\examples{
# load data
data(trawl_fishery_Java)

# run model
output = SchaeferFox(Y = trawl_fishery_Java$yield,
                     f = trawl_fishery_Java$effort)
output[[2]]

}
\references{
Fox, W. W. Jr., 1970. An exponential surplus-yield model for optimizing exploited fish
populations. \emph{Trans.Am.Fish.Soc.}, 99:80-88

Graham, M., 1935. Modern theory of exploiting a fishery and application to North Sea
trawling. \emph{J.Cons.CIEM}, 10(3):264-274

Schaefer, M., 1954. Some aspects of the dynamics of populations important to the
management of the commercial marine fisheries. \emph{Bull.I-ATTC/Bol. CIAT}, 1(2):27-56

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

