% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Selectivity model}
\usage{
select(param)
}
\arguments{
\item{param}{A list with following parameters: midlengths of size classes
(\code{$midLengths}), number of fish caught with net 1 (\code{$numNet1}),
number of fish caught with net 2 (\code{$numNet2}), and the meshsizes of both
nets (\code{$msNet1} & \code{$msNet2}).}
}
\description{
This function estimates the selecitvity of different gillnet and trawl nets
   from experimental catches.
}
\details{
This function estimates the fractions retained by each net (SNet1 & SNet2), the
  optimum lengths for each net, the selection factor (SF), and the standard deviation
  of the factor (stand.dev). Calculations are based on a normal distribution with common spread.
  Assumptions of this method are, that (i) the optimum length Lm is proportional to the mesh
  size (Lm = SF * m), (ii) the selection curves are normally distributed with a common
  standard deviation, (iii) the nets have the same fishing power (same dimensions and material).
  Requirements for the experimental set-up are: selection curves corresponding to the two
  mesh sizes have to overlap, and the nets have to be set in the same area, during the
  same time.
  To calculate selection factor (SF), L25, L50 and L75 for trawl nets /fisheries.
}
\examples{
# Gillnet selectivity
# load data
data(tilapia)              ### == data_GillnetSelect

# run model
output <- select(tilapia)

# plot results
plot(output)

# Trawl selectivity
# load data
data(bream)                #### == data_TrawlSelect

# run model
output <- select(bream)

# plot results
plot(output)

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

