% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_Millar.R
\name{select_Millar}
\alias{select_Millar}
\title{Millar's selectivity model}
\source{
https://www.stat.auckland.ac.nz/~millar/selectware/
}
\usage{
select_Millar(param, x0, rtype = "norm.loc", rel.power = NULL)
}
\arguments{
\item{param}{A list with following parameters: vector with midlengths of size classes
(\code{$midLengths}), vector with meshSizes in increasing order (\code{$meshSizes}),
and a matrix with the number of individuals caught with each sized mesh
(\code{$CatchPerNet_mat}).}

\item{model}{A character string indicating which type fo model for the estimation
of the selection curves should be used: \code{"fixed_normal"}, \code{"normal"},
\code{"gamma"}, or \code{"lognormal"}.}

\item{plotlens}{Vector of lengths for which values of relative retention are required}

\item{rel}{Vector of realtive efficencies}
}
\description{
This function estimates the selecitvity of a gillnet from an experimental
   catch with two gillnets with different mesh sizes.
}
\details{
Model adapted from the selectivity functions provided by Prof. Dr. Russell Millar
  (https://www.stat.auckland.ac.nz/~millar/).

  This function estimates the fractions retained by each net (SNet1 and SNet2), the
  optimum lengths for each net, the selection factor (SF), and the standard deviation
  of the factor (stand.dev).
  Assumptions of this method are, that (i) the optimum length Lm is proportional to the mesh
  size (Lm equals SF times m), (ii) the selection curves are normally distributed with a common
  standard deviation, (iii) the nets have the same fishing power (same dimensions and material).
  Requirements for the experimental set-up are: selection curves corresponding to the two
  mesh sizes have to overlap, and the nets have to be set in the same area, during the
  same time.
}
\examples{
# Trouser trawl net
# load data
data(haddock)

# run model
output <- select_Millar(haddock, x0 = c(-10,0.3,0),
   rtype = "tt.logistic", rel.power = NULL)

# plot results
plot(output, plotlens=seq(25,35,0.1))
legend(32,0.5,c("Control","Experimental"),lty=1:2,col=1:2,cex=0.8,bg="white")


# Gillnet
# load data
data(gillnet)

# run model
output <- select_Millar(gillnet, x0 = c(60,4), rel.power = rep(1,8),
   rtype = "norm.loc")

# investigate results
output

# plot results
plot(output,plotlens=seq(40,90,0.1))


# Stacked trammel net
#Here, the data are from two expts using different mesh sizes
#This analysis assumes common retention curve in both expts.
#Note that summary function does not produce residual plot
#since lengths are not unique
# load data
data(trammelnet)

# run model
output <- select_Millar(trammelnet, x0 = c(25,4),
   rtype="norm.loc", rel.power = rep(1,6))

# plot results
plot(output,plotlens=seq(10,40,0.1))


}
\references{
Millar, R. B., Holst, R., 1997. Estimation of gillnet and hook selectivity
 using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.

 Holt, S. J. 1963. A method for determining gear selectivity and its application.
 ICNAF Special Publication, 5: 106-115.
}

