% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ogive.R
\name{select_ogive}
\alias{select_ogive}
\title{Selectivity patterns}
\usage{
select_ogive(s_list, Lt, Lc = NA)
}
\arguments{
\item{s_list}{a list with selectivity parameters dependent on the type of
   selection curve:
\itemize{
  \item \code{selecType} type of
   selection curve used for estimation (options:
   "knife_edge",
   "trawl_ogive",
   "lognormal",
   "normal_fixed"),
  \item \code{Lc} length-at-first-capture,
  \item \code{meshSizes} a vector with mesh sizes in increasing order,
  \item \code{select_ps} selectivity parameters,
  \item \code{L75} length at which individuals are caught with a
   probability of 75%.
}}

\item{Lt}{a vector with lengths corresponding to age classes}

\item{Lc}{length-at-first-capture (Default: NA)}
}
\description{
Based on a few parameters, this function estimates the fraction per
   length group retained in the net. Different selection curves can be used for the
   estimation.
}
\details{
This function is embedded within \code{\link{predict_mod}}. \code{selecType}
   "knife_edge" only requires a Lc value. "trawl_ogive" requires a Lc (L50) and
   a L75 value. "lognormal" and "normal_fixed" require two mesh sizes with two
   corresponding "select_ps" values.
}
\examples{
# create list with selectivity information
select.list <- list(selecType = 'knife_edge',
   Lc = 34, tc = 5, meshSizes = c(8.1,9.1),
   select_ps = c(21.1,23.8))

# create vector with mid lengths
Lt <- c(12,22,29,34,38,41,43,45,46,47,48,49,49)

# run model
select_ogive(select.list, Lt)

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.
}

