% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrimps.R
\docType{data}
\name{shrimps}
\alias{shrimps}
\title{Shrimp data}
\format{
A list consisting of: 1. age groups [months], 2. meanWeight, 3. meanValue,
   4.FM, and 5.M
}
\source{
Garcia, S. and N.P. van Zalinge, 1982. Shrimp fishing in Kuwait: methodology
   for a joint analysis of the artisanal and industrial fisheries. pp. 119-142 In:
   Report on the Workshop on assessment of the shrimp stocks of the west coast of
   the Gulf between Iran and the Arabian Peninsula. Fisheries development in the
   Gulf. Rome, FAO, FI:DP/RAB/80/015/1, 163 p.
}
\usage{
data(shrimps)
}
\description{
Data from the Kuwait shrimp (\emph{Penaeus semisulcatus}) fishery.
   They can be used to simulate different management scenarios by applying
   prediction models (\code{\link{predict_mod}}).
   Fishing mortalities were estimated from catch data and biomass estimates were
   obtained from the swept area method. Recuritment to fishery at the age of one month (Tr = 1).
}
\examples{
data(shrimps)
str(shrimps)
summary(shrimps)


}
\keyword{data}
\keyword{dataset}
\keyword{length}
\keyword{prediction}
