% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synCAA.R
\docType{data}
\name{synCAA}
\alias{synCAA}
\title{Synthetic Catch-at-age data from Sparre & Venema (1998)}
\format{A list consisting of:
\itemize{
  \item \code{$midAge} a vector of the mid ages of the age groups,
  \item \code{$catch_mat} a matrix with the catches for different years.
}}
\source{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment. Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\usage{
data(synCAA)
}
\description{
Synthetic Catch-at-age data. Can be used for the calculation of Z by means of
   \code{\link{BevertonHoltsZ}}.
}
\examples{
data(synCAA)
str(synCAA)


}
\keyword{CAA}
\keyword{Catch-at-age}
\keyword{age}
\keyword{data}
\keyword{dataset}

