% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synLFQ3.R
\docType{data}
\name{synLFQ3}
\alias{synLFQ3}
\title{Synthetic length frequency data III from Sparre & Venema (1998)}
\format{A list consisting of:
\itemize{
  \item \strong{midLengths}: a vector of the mid lengths of the length groups,
  \item \strong{Linf}: infinite length for investigated species in cm [cm],
  \item \strong{K}: growth coefficent for investigated species per year [1/year],
  \item \strong{t0}: theoretical time zero, at which individuals of this species hatch,
  \item \strong{catch}: a vector with catches.
}}
\source{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment. Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\usage{
data(synLFQ3)
}
\description{
Synthetic length frequency data III. Can be used for the calculation of Z
   by means of \code{\link{Z_BevertonHolt}}.
}
\examples{
data(synLFQ3)
str(synLFQ3)


}
\keyword{data}
\keyword{length-frequency}

