% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_anchor2t0.R
\name{t_anchor2t0}
\alias{t_anchor2t0}
\title{Calculate age-based VBGF parameters from time-based estimates}
\usage{
t_anchor2t0(par = NULL, Lrecr = NULL, plot = TRUE)
}
\arguments{
\item{par}{list. VBGF parameters.}

\item{Lrecr}{numeric. Length at recruitment
(default: \code{Lrecr = NULL}).}

\item{plot}{logical. Plot graphical representation of both
time and age based growth curves (default: \code{plot = TRUE}).}
}
\value{
a list containging age-based VBGF parameters.
}
\description{
Translates t_anchor to t0 and ts to age-adjusted ts. In addition to
VBGF parameters calculated via e.g. ELEFAN, the length at recruitment
(\code{Lrecr}) must also be provided.
}
\examples{
lfq <- synLFQ4
lfq$par <- list(
  Linf = 80, K = 0.5, t_anchor = 0.25, C = 0.75, ts = 0.5
)
t_anchor2t0(par = lfq$par, Lrecr = 10)

}
