% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ta2t0}
\alias{ta2t0}
\title{Calculate age-based VBGF parameters from time-based estimates}
\usage{
ta2t0(par = NULL, L0 = 0, plot = TRUE)
}
\arguments{
\item{par}{list. VBGF parameters.}

\item{L0}{numeric. Length at age zero (i.e. recruitment)
(default: \code{L0 = 0}).}

\item{plot}{logical. Plot graphical representation of both
time and age based growth curves (default: \code{plot = TRUE}).}
}
\value{
a list containging age-based VBGF parameters.
}
\description{
Translates ta to t0 and ts to age-adjusted ts. In addition to
VBGF parameters calculated via e.g. ELEFAN, the length at recruitment
(\code{L0}) must also be provided.
}
\examples{
data(synLFQ4)
lfq <- synLFQ4
lfq$par <- list(
  Linf = 80, K = 0.5, ta = 0.25, C = 0.75, ts = 0.5
)
ta2t0(par = lfq$par, L0=10)

}
