% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{A test data set}
\format{A dataframe consisting of: 1.bla, 2. bla, and 3.
bla (10 rows, 3 columns)}
\source{
\url{http://www.esrl.noaa.gov/psd/data/gridded/data.kaplan_sst.html}
}
\usage{
data(testdata)
}
\description{
The \code{testdata} data set is bull

\itemize{
  \item N. bla1
  \item L. bla2
  \item TYPE. bla3
}
}
\examples{
### Ex 1.
data(testdata)
plot(testdata)
}
\keyword{datasets}

