% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmtool.R
\name{TFSA2DLMtool}
\alias{TFSA2DLMtool}
\title{Harvest control rules based on TropFishR}
\usage{
TFSA2DLMtool(uncertaintyCap = FALSE, lower = 0.8, upper = 1.2,
  env = globalenv())
}
\arguments{
\item{uncertaintyCap}{Logical; If true TAC is bound between two values set in lower and upper. Default: FALSE.}

\item{lower}{lower bound of the uncertainty cap. Default is 0.8, used if uncertaintyCap = TRUE.}

\item{upper}{upper bound of the uncertainty cap. Default is 1.2, used if uncertaintyCap = TRUE.}
}
\value{
A function which can estimate TAC recommendations based on SPiCT assessment,
  taking assessment uncertainty into account.
}
\description{
This function creates an harvest control rule for use with the
  DLMtool package.
}
\examples{
\dontrun{
library(DLMtool)

## Put together an operating model from the available DLM toolkit examples
StockEx <- Red_snapper
FleetEx <- Generic_IncE
ObsEx <- Precise_Unbiased

## Remove changes in life history parameters
StockEx@Mgrad <- c(0,0)
StockEx@Kgrad <- c(0,0)
StockEx@Linfgrad <- c(0,0)
StockEx@Prob_staying <- c(1,1)

## Set the depletion level 
StockEx@D <- c(0.3, 0.4)

## create Operation Model
OMex <- new("OM", Stock = StockEx, Fleet = FleetEx, 
                  Obs = ObsEx)

## Set simulation options
OMex@nsim <- 10
OMex@nyears <- 25
OMex@proyears <- 5

## Get TFSA HCR
MPname <- TFR2DLMtool(uncertaintyCap = c(FALSE,TRUE))
MPname

## run MSE
MSEex <- runMSE(OMex, MPs = MPname, CheckMPs = FALSE)
Pplot2(MSEex, traj="quant", quants=c(0.2, 0.8))
}


}
