% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawl_fishery_Java.R
\docType{data}
\name{trawl_fishery_Java}
\alias{trawl_fishery_Java}
\title{Data from the trawl fishery off the North coast of Java}
\format{A dataframe consisting of: 1. \strong{year} a vector with years, 2. \strong{Y}
 yield, and 3. \strong{f} fishing effort.}
\source{
Dwiponggo, A., 1979. Review of the demersal resources and fisheries in the
 Java Sea. IPFC:RRD/II/79/Inf.12. Paper presented at the SCORRAD Meeting, 1979,
 Hong Kong.
}
\usage{
data(trawl_fishery_Java)
}
\description{
Times series of catch and effort data from the trawl fishery off the
 North coast of Java. This dataset can
 be used for the calculation procedure of the Schaefer and Fox model
 (\code{\link{SchaeferFox}}).
}
\examples{

data(trawl_fishery_Java)
str(trawl_fishery_Java)


}
\keyword{Java}
\keyword{dataset}
\keyword{trawl}

