% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrawlSelect.R, R/TrawlSelect2.R
\name{TrawlSelect}
\alias{TrawlSelect}
\title{Trawl net selctivity}
\usage{
TrawlSelect(param = data_TrawlSelect)

TrawlSelect(param = data_TrawlSelect)
}
\arguments{
\item{param}{List with parameters: midlengths of size classes, number in net, number in codend and meshsize of codend}

\item{classes}{Midpoints of the length class as vector (length frequency data) or ages as vector (age composition data).}

\item{catch}{Catch as vector, or a matrix with catches of subsequent years if the catch curve with constat time intervals should be applied.}

\item{datatype}{Type of data which is used for analysis, either 'length' or 'age', for length frequency or age composition data, respectively}

\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficent for investigated species per year [1/year].}

\item{t0}{Theoretical time zero, at which individuals of this species hatch (default: 0).}
}
\description{
This function estimates the selecitvity of trawl nets.

This function estimates the selecitvity of trawl nets.
}
\details{
To calculate selection factor (SF), L25, L50 and L75 for trawl nets /fisheries.

To calculate selection factor (SF), L25, L50 and L75 for trawl nets /fisheries.
}
\examples{
\donttest{
data("ex.TrawlSelect")
output <- with(ex.TrawlSelect,TrawlSelect(classes = midLengths,
   numCodend, numCover, meshsizeCodend = 4))
output
}

data(data_TrawlSelect)
output <- with(data_TrawlSelect,TrawlSelect())
plot(output)

}
\references{
xx

xx
}

