% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trout.R
\docType{data}
\name{trout}
\alias{trout}
\title{Trout data}
\format{A list consisting of: 1. sample numbers, 2. the midlengths of size classes,
 2. a matrix with catch in numbers.}
\source{
Goeden, G.B., 1978. A monograph of the coral trout \emph{Plectropomus leopardus}
(Lacepede). \emph{Res.Bull.Fish.Serv.Queensl.}, (1):42 p.

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.
}
\usage{
data(trout)
}
\description{
This dataset contains information about coral trout
 (\emph{Plectropomus leopardus}) caught near Heron Island (Great Barrier Reef,
 Australia) in October 1971.
 It can be used for the estimation of growth parameters, \code{\link{ELEFAN}}.
}
\examples{

data(trout)
str(trout)
summary(trout)


}
\keyword{data}
\keyword{dataset}
\keyword{length-frequency}
\keyword{trout}

