% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBGF.R
\name{VBGF}
\alias{VBGF}
\title{Von Bertalanffy Growth function (VBGF)}
\usage{
VBGF(t, Linf = NA, Winf = NA, K, t0 = 0, b = 3, D = 1, L0 = NA,
  ts = 0, C = 0)
}
\arguments{
\item{t}{age}

\item{Linf}{infinite length for investigated species in cm, or}

\item{Winf}{infinite weight for investigated species in gramm}

\item{K}{growth coefficent for investigated species per year}

\item{t0}{theoretical time zero, at which individuals of this species hatch}

\item{b}{exponent of weight length relationship}

\item{D}{surface factor}

\item{L0}{length at hatching for VBGF with L0}

\item{ts}{onset of the first oscillation relative to t0}

\item{C}{intensity of (sinusoid) growth oscillations}
}
\value{
A vector with estimated lengths corresponding to provided ages.
}
\description{
This function applies the von Bertalanffy growth function. It
   allows to calculate ages from lengths based on the special, generalised
   or seasonalised VBGF.
}
\details{
Based upon which input parameters are given one of the following
VBGF types is applied: "special", "generalised", or "seasonalised" VBGF.
}
\examples{
# with t0
t <- seq(0,6,0.1)
Lt <- VBGF(Linf=80, K=0.6, t=t, t0=-0.1)
plot(t, Lt, t="l")

# with L0
t <- seq(0,6,0.1)
Lt <- VBGF(Linf=80, K=0.6, t=t, L0=2)
plot(t, Lt, t="l")

#seasonalised
t <- seq(0,6,0.1)
Lt <- VBGF(Linf=80, K=0.6, t=t, t0=-0.1, ts=0.5, C=0.75)
plot(t, Lt, t="l")

}
\references{
Somers, I. F. (1988). On a seasonally oscillating growth function.
Fishbyte, 6(1), 8-11

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.
}
\keyword{VBGF}
\keyword{function}
\keyword{growth}

