% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFANboot.R
\name{vbgfCI_time}
\alias{vbgfCI_time}
\title{VBGF plot and CI}
\usage{
vbgfCI_time(res, CI = 95, agemax = NULL, add_legend = TRUE,
  add_max_dens_legend = TRUE, xlab = "Relative time",
  ylab = "Length", perm.col = adjustcolor("grey50", 0.1),
  perm.lwd = 1, ci.col = 1, ci.lty = 2, ci.lwd = 1, maxd.col = 1,
  maxd.lty = 1, maxd.lwd = 2)
}
\arguments{
\item{res}{list of class `lfqBoot`}

\item{CI}{vector. Confidence interval levels to plot}

\item{agemax}{numeric. Maximum number of years to project.}

\item{add_legend}{logical. Should CI and max. density legend be added
(Default: `add_legend = TRUE`).}

\item{add_max_dens_legend}{logical. Should maximum density line be added
(Default: `add_max_dens_legend = TRUE`).}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{perm.col}{Color for each resample estimate line. See `?par`.}

\item{perm.lwd}{Line width for each resample estimate line. See `?par`.}

\item{ci.col}{Color for CI line. See `?par`.}

\item{ci.lty}{Line type for CI line. See `?par`.}

\item{ci.lwd}{Line width for CI line. See `?par`.}

\item{maxd.col}{Color for maximum density line. See `?par`.}

\item{maxd.lty}{Line type maximum density line. See `?par`.}

\item{maxd.lwd}{Line width maximum density line. See `?par`.}
}
\value{
plot and list containing: `limCI` - data.frame with CI limits by time,
  `inCi` -  data.frame with logical values defining whether bootstrapping samples
  are within each of the defined CIs, `density` - the multivariate kernel density
  estimates for each sample, and `max_dens` is a list with the VBGF parameter
  combination having the maximum density estimate.
}
\description{
VBGF plot and CI
}
\examples{

data(alba_boot)
CIinfo <- vbgfCI_time(
  res = alba_boot,
  agemax = 2, CI = 95,
  perm.col = adjustcolor("grey50",0.2)
)

# plot more CI levels
CIinfo <- vbgfCI_time(
  res = alba_boot,
  agemax = 2, CI = c(95, 50),
  ci.lty = 1, ci.lwd = 2, ci.col = c("red", "orange"),
  perm.col = adjustcolor("grey50",0.2)
)

# using output in lfq plot (see ?TropFishR::plot.lfq)
library(TropFishR)
data(alba)
alba <- lfqRestructure(alba, MA = 7)
plot(alba, Fname = "rcounts")
for(i in seq(nrow(alba_boot$bootRaw))){
  x <- as.list(alba_boot$bootRaw[i,])
  tmp <- lfqFitCurves(
    lfq = alba, par = x,
    col = adjustcolor("grey50",0.1), draw = TRUE, lty=1
  )
}
tmp <- lfqFitCurves(lfq = alba, par = CIinfo$max_dens,
  col = 1, draw = TRUE, lty=1, lwd=2
)

}
