% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bestCC}
\alias{bestCC}
\title{Automated determination of values to include in catch curve}
\usage{
bestCC(x, y, minN = 6)
}
\arguments{
\item{x}{vector. Age or relative age associated with catch numbers
(\code{x}).}

\item{y}{vector. catch numbers.}

\item{minN}{numeric. Minimum number of values to include in the regression
(Default = 3).}
}
\value{
list containing the indices and F-statistic of the highest scoring
subset of (sequantial) data points.
}
\description{
Procedure described by Pauly (1990) to automate the selection
  of points used in a catch curve (\code{log(y)~x}). The sequential
  data indices that result in the highest F-ststistic in a linear regression
  are returned. Results should still be visualized to prevent suboptimal
  results form 'pathological' datasets.
}
\examples{
# generate example data
set.seed(1)
n <- 10
x <- seq(n)
logy <- 10 + -2*x + rnorm(n, sd = 0.5)
logy[1:2] <- logy[1:2] * c(0.5, 0.75)
y <- exp(logy)
plot(y ~ x, log = "y")

# When lm can reduce to 3 points, this is chosen as the 'best'
(tmp <- bestCC(x = x, y = y, minN = 3))
plot(y ~ x, log = "y")
points(x[tmp$samples], y[tmp$samples], pch = 16)

# Any min number of points above this threshhold increases the best n a lot
(tmp <- bestCC(x = x, y = y, minN = 4))
plot(y ~ x, log = "y")
points(x[tmp$samples], y[tmp$samples], pch = 16)

}
\references{
Pauly, D. (1990). Length-converted catch curves and the seasonal growth
  of fishes. Fishbyte, 8(3), 33–38.
}
