% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BevertonHoltsZ.R
\name{BevertonHoltsZ}
\alias{BevertonHoltsZ}
\title{Beverton and Holt's Z-Equations}
\usage{
BevertonHoltsZ(param, catch_column = NA, PowellWetherall = FALSE)
}
\arguments{
\item{param}{A list consisting of following parameters:
\code{$midAge} or \code{$midLengths} ages as vector (age composition data) or
midpoints of the length class as vector (length frequency data),
\code{$Linf} Infinite length for investigated species in cm [cm],
\code{$K} Growth coefficent for investigated species per year [1/year],
\code{t0} Theoretical time zero, at which individuals of this species hatch,
\code{catch} or \code{catch_mat} Catch as vector, or a matrix with catches of subsequent years if
the catch curve with constat time intervals should be applied;}

\item{catch_column}{A Number indicating which column of \code{catch_mat} should be analysed (Default: \code{catch_column = NA}).}

\item{PowellWetherall}{A logical parameter indicating if the Powell Wetherall method should be applied to length frequency data (Default: \code{PowellWetherall = FALSE})}
}
\description{

}
\details{
Lprime or tprime will be identified via the first length (or age) class inserted.For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
# Length based model
# load data
data(synLFQ2)

# run model
BevertonHoltsZ(synLFQ2, catch_column = 2)

# Age based model
# load data
data(synCAA)

#run model
BevertonHoltsZ(synCAA, catch_column = 3)

\donttest{

# load data
data(synLFQ3)

# run model
BevertonHoltsZ(synLFQ3,PowellWetherall = T)
}

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

