% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BevertonHoltsZ.R
\name{BevertonHoltsZ}
\alias{BevertonHoltsZ}
\title{Beverton and Holt's Z-Equations}
\usage{
BevertonHoltsZ(classes, catch, datatype, PowellWetherall = FALSE,
  Linf = NULL, K = NULL)
}
\arguments{
\item{classes}{Midpoints of the length class as vector (length frequency data) or ages as vector (age composition data).}

\item{catch}{Catch as vector, or a matrix with catches of subsequent years if the catch curve with constat time intervals should be applied.}

\item{datatype}{Type of data which is used for analysis, either 'length' or 'age', for length frequency or age composition data, respectively}

\item{PowellWetherall}{A logical parameter indicating if the Powell Wetherall method should be applied to length frequency data.}

\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficent for investigated species per year [1/year].}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve or from age at length data with catch curve.
}
\details{
Lprime or tprime will be identified via the first length (or age) class inserted.For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
# load data
data("ex.BevertonHoltsZ")

# run model
with(ex.BevertonHoltsZ, BevertonHoltsZ(midLength,
 catch1960, datatype="length", Linf = 100, K = 0.3))

# load data
data("ex.ageBevertonHoltsZ")

#run model
with(ex.ageBevertonHoltsZ,BevertonHoltsZ(midAge,
 catch1960,datatype = 'age'))

\donttest{

# load data
data("ex.PowellWetherall")

# run model
with(ex.PowellWetherall, BevertonHoltsZ(midLength,
  catch, datatype="length", PowellWetherall = T))

# load data
data("ex.PowellWetherall2")

# run model
with(ex.PowellWetherall2, BevertonHoltsZ(midLength,
  catch, datatype="length", PowellWetherall = T))
}

}
\references{
Sparre ???
}

