% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bhattacharya.R
\name{Bhattacharya}
\alias{Bhattacharya}
\title{Bhattacharya's method}
\usage{
Bhattacharya(param, n_rnorm = 1000)
}
\arguments{
\item{param}{a list consisting of following parameters:
\code{midLengths} midpoints of the length class as vector,
\code{catch} a vector with the catch per length class or a matrix with
   catches per length class of subsequent years;}

\item{n_rnorm}{number of observations for the function \code{\link{rnorm}}, default = 1000}
}
\value{
A list with the input parameters and a data.frame (\strong{bhat_results})
     with the results of the Bhattacharya method. Most important is the column
     \strong{Nx}, where x is the number of the cohort, displaying the number of
     individuals per cohort. For an explanation of the other columns please refer to
     the FAO manual (Sparre & Venema, 1998).
}
\description{
Find relative frequencies and frequency distribution of distinct cohorts
     in the observed length frequency distribution by resolving it into
     Gaussian components.
}
\details{
This method includes the \code{\link{identify}} function, which allows to
     choose points on a graphical device manually. To stop this process please press
     right mouse click on the graph, and in case of windows click on "Stop".
}
\examples{
\donttest{
 data(synLFQ1)
 Bhattacharya(synLFQ1)
}

}
\references{
Bhattacharya, C.G., 1967. A simple method of resolution of a distribution into Gaussian
components, \emph{Biometrics}, 23:115-135

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.
}
\keyword{Bhattacharya}
\keyword{length-frequency}

