% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bhattacharya.R
\name{Bhattacharya}
\alias{Bhattacharya}
\title{Bhattacharya's method}
\usage{
Bhattacharya(midLengths, catch)
}
\arguments{
\item{midLengths}{Midpoints of the length class as vector}

\item{catch}{Catch per sampling time as matrix or the total catch as vector.}
}
\description{
Separating normal distributions of several cohorts in distinct
  distributions representing cohorts. Second sentence.
}
\details{
Bhattacharya
}
\examples{
\donttest{
 data("ex.Bhattacharya")
 output = with(Bhattacharya(midLengths,catch))
 output
}
}
\references{
Jones Sparre
}

