% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CatchCurve.R, R/CatchCurvePlusSelectOgive.R
\name{CatchCurve}
\alias{CatchCurve}
\title{Catch curve}
\usage{
CatchCurve(classes, catch, datatype, Linf = NULL, K = NULL, t0 = 0)

CatchCurve(classes, catch, datatype, Linf = NULL, K = NULL, t0 = 0)
}
\arguments{
\item{classes}{Midpoints of the length class as vector (length frequency data) or ages as vector (age composition data).}

\item{catch}{Catch as vector, or a matrix with catches of subsequent years if the catch curve with constat time intervals should be applied.}

\item{datatype}{Type of data which is used for analysis, either 'length' or 'age', for length frequency or age composition data, respectively}

\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficent for investigated species per year [1/year].}

\item{t0}{Theoretical time zero, at which individuals of this species hatch (default: 0).}

\item{classes}{Midpoints of the length class as vector (length frequency data) or ages as vector (age composition data).}

\item{catch}{Catch as vector, or a matrix with catches of subsequent years if the catch curve with constat time intervals should be applied.}

\item{datatype}{Type of data which is used for analysis, either 'length' or 'age', for length frequency or age composition data, respectively}

\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficent for investigated species per year [1/year].}

\item{t0}{Theoretical time zero, at which individuals of this species hatch (default: 0).}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve or from age at length data with catch curve.

This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve or from age at length data with catch curve.
}
\details{
For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.

For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{
\donttest{
# Variable paramter system
# based on length frequency data
data("ex.LengthCC")
output <- with(ex.LengthCC,CatchCurve(ex.LengthCC[,1], ex.LengthCC[,2],
Linf = 23.1, K = 0.59, datatype = 'length')
output
# based on age composition data
data("ex.CatchCurve")
output <- with(ex.CatchCurve,CatchCurve(age, ex.CatchCurve[,2],datatype = 'age'))
output
# Constant parameter system based on age composition data
output <- with(ex.CatchCurve,CatchCurve(age, ex.CatchCurve[,2:8],datatype = 'age'))
output
 }
\donttest{
# Variable paramter system
# based on length frequency data
data("ex.LengthCC")
output <- with(ex.LengthCC,CatchCurve(ex.LengthCC[,1], ex.LengthCC[,2],
Linf = 23.1, K = 0.59, datatype = 'length')
output
# based on age composition data
data("ex.CatchCurve")
output <- with(ex.CatchCurve,CatchCurve(age, ex.CatchCurve[,2],datatype = 'age'))
output
# Constant parameter system based on age composition data
output <- with(ex.CatchCurve,CatchCurve(age, ex.CatchCurve[,2:8],datatype = 'age'))
output
 }
}
\references{
Sparre ??? data from ICES 1981a

Sparre ??? data from ICES 1981a
}

