% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFAN.R
\name{ELEFAN}
\alias{ELEFAN}
\title{ELEFAN}
\usage{
ELEFAN(param, range.Linf, step.Linf, range.K, step.K, t0 = 0, tmax,
  growth.curve = "VBGF", seed.C = 0, seed.WP = 0, conf.int = FALSE)
}
\arguments{
\item{param}{A list consisting of following parameters:
\itemize{
  \item \strong{age} or \strong{midLengths} midpoints of the length class as vector (length frequency
  data) or ages as vector (age composition data),
  \item \strong{catch} Catch as vector, or a matrix with catches of subsequent years if
  the catch curve with constat time intervals should be applied;
}}

\item{range.Linf}{lower and upper limit of range of Linf values}

\item{step.Linf}{step size between Linf values}

\item{range.K}{lower and upper limit of range of K values}

\item{step.K}{step size between K values}

\item{t0}{theoretical age of fish at length zero}

\item{tmax}{maximal age of fish from literature only estimate needed}

\item{conf.int}{logical; if \code{TRUE} calculates the 95% confidence intervals for
calculated Linf and K parameters by means of the jackknife method (Quenouille, 1956; Tukey, 1958, 1986),}

\item{C}{growth oscillation amplitude}

\item{WP}{winter point}
}
\description{
Electronic length frequency analysis for estimating growth parameter of exploited fish populations.
}
\details{
Winter point (WP) designates the period of the year when growth is slowest, which
   in the northern hemisphere is often found around 0.2 (February) and in the southern
   hemisphere aorund 0.7 (October). C express the amplitude of growth oscillations and
   should be between 0 (no oscillation) and 1 (no growth), values above 1 suggest
   a reduction of length and thus are unlikely (Pauly and Morgan, 1987).
}
\examples{
\donttest{
data(trout)
ELEFAN(trout, range.Linf = c(11,15), step.Linf = 1, range.K = c(0.2,1),
   step.K = 0.1, tmax = 5)

param = trout
range.Linf = c(11,15)
step.Linf = 1
range.K = c(0.2,1)
step.K = 0.1
t0 = 0
C = 0
WP = 0
growth.curve = "VBGF"
tmax = 5
conf.int = FALSE
}

}
\references{
Pauly, D. 1981. The relationship between gill surface area and growth performance in fish:
a generalization of von Bertalanffy's theory of growth. \emph{Meeresforsch}. 28:205-211

Pauly, D. and N. David, 1981. ELEFAN I, a BASIC program for the objective extraction of
growth parameters from length-frequency data. \emph{Meeresforschung}, 28(4):205-211

Pauly, D., 1985. On improving operation and use of ELEFAN programs. Part I: Avoiding
"drift" of K towards low values. \emph{ICLARM Conf. Proc.}, 13-14

Pauly, D., 1987. A review of the ELEFAN system for analysis of length-frequency data in
fish and aquatic invertebrates. \emph{ICLARM Conf. Proc.}, (13):7-34

Pauly, D. and G.R.Morgan (Eds.), 1987. Length-based methods in fisheries research.
(No. 13). WorldFish

Quenouille, M. H., 1956. Notes on bias in estimation. \emph{Biometrika}, 43:353-360

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2): 407 p.

Tukey, J., 1958. Bias and confidence in not quite large samples.
\emph{Annals of Mathematical Statistics}, 29: 614

Tukey, J., 1986. The future of processes of data analysis. In L. V. Jones (Eds.),
The Collected Works of John W. Tukey- philosophy and principles of data analysis:
1965-1986 (Vol. 4, pp. 517-549). Monterey, CA, USA: Wadsworth & Brooks/Cole
}

