% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFAN_SA.R
\name{ELEFAN_SA}
\alias{ELEFAN_SA}
\title{ELEFAN_SA}
\usage{
ELEFAN_SA(x, seasonalised = FALSE, init_par = list(Linf = 50, K = 0.5,
  t_anchor = 0.5, C = 0, ts = 0), low_par = NULL, up_par = NULL,
  SA_time = 60 * 3, SA_temp = 1e+05, verbose = TRUE, MA = 5,
  addl.sqrt = FALSE, flagging.out = TRUE, plot = FALSE)
}
\arguments{
\item{x}{a list consisting of following parameters:
\itemize{
  \item \strong{midLengths} midpoints of the length classes,
  \item \strong{dates} dates of sampling times (class Date),
  \item \strong{catch} matrix with catches/counts per length class (row) and sampling date (column);
}}

\item{seasonalised}{logical; FALSE}

\item{init_par}{default: list(Linf = 50, K = 0.5, t_anchor = 0.5, C = 0, ts = 0)}

\item{low_par}{list, default NULL, e.g. list(Linf = 1, K = 0.01, t_anchor = 0, C = 0, ts = 0)}

\item{up_par}{list, default NULL, e.g. list(Linf = 1000, K = 10, t_anchor = 1, C = 1, ts = 1)}

\item{SA_time}{default = 60 * 1}

\item{SA_temp}{default = 100000}

\item{verbose}{default   = TRUE}

\item{MA}{number indicating over how many length classes the moving average should be performed (defalut: 5, for
more information see \link{lfqRestructure}).}

\item{addl.sqrt}{Passed to \link{lfqRestructure}. Applied an additional square-root transformation of positive values according to Brey et al. (1988).
(default: FALSE, for more information see \link{lfqRestructure}).}

\item{flagging.out}{logical; passed to \link{calcLt}. Default is TRUE}

\item{plot}{logical; indicating if plot with restrucutred frequencies and growth curves should
be displayed}
}
\value{
A list with the input parameters and following list objects:
\itemize{
  \item \strong{samplingPeriod}: length of sampling period in years,
  \item \strong{samplingDays}: time of sampling times in relation to first sampling time,
  \item \strong{delta_t}: array with time differences between relative sampling time set to zero and
     other sampling times,
  \item \strong{rcounts}: restructured frequencies,
  \item \strong{peaks_mat}: matrix with positive peaks with distinct values,
  \item \strong{ASP}: available sum of peaks, sum of posititve peaks which could be potential be hit by
     growth curves,
  \item \strong{C}: amplitude of growth oscillation,
  \item \strong{ts}: winter point winter point (WP = ts + 0.5);
when the K-Scan method is applied (fixed Linf) in addition following parameters:
  \item \strong{Rn_max}: highest score value,
}
}
\description{
Electronic LEngth Frequency ANalysis with simulated annealing
   for estimating growth parameters.
}
\details{
This is cool function
}
\examples{
\dontrun{

## trout example
data(trout)

# ELEFAN_SA (takes approximately 10 seconds)
output <- ELEFAN_SA(trout, SA_time = 30)
output$par
output$Rn_max

# view fit
plot(output, ylim = c(0,15))
calcLt(output, col=1, par=output$par, draw=TRUE)$ESP



## synthetic lfq example
data(synLFQ4)

# ELEFAN_SA (takes approximately 2 minutes)
output <- ELEFAN_SA(synLFQ4, SA_time = 60*2, seasonalised = TRUE, MA = 15,
  init_par = list(Linf = 75, K = 0.5, t_anchor = 0.5, C = 0.5, ts = 0.5),
  low_par = list(Linf = 70, K = 0.3, t_anchor = 0, C = 0, ts = 0),
  up_par = list(Linf = 90, K = 0.7, t_anchor = 1, C = 1, ts = 1),
)
output$par
output$Rn_max

# view fit
plot(output)
calcLt(output, col=1, par=output$par, draw=TRUE)$ESP

# compare to original parameters
tmp <- calcLt(
  output, col=4, lty=1,
  par=list(Linf=80, K=0.5, t_anchor=0.25, C=0.75, ts=0),
  draw=TRUE
)
tmp$ESP
output$Rn_max

}


}
\references{
Pauly, D. and N. David, 1981. ELEFAN I, a BASIC program for the objective extraction of
growth parameters from length-frequency data. \emph{Meeresforschung}, 28(4):205-211
}

