% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.ageBevertonHoltsZ.R
\docType{data}
\name{ex.ageBevertonHoltsZ}
\alias{ex.ageBevertonHoltsZ}
\title{Example dataset of length for use in BevertonHoltsZ}
\format{A list consisting of: 1. a da}
\usage{
data(ex.ageBevertonHoltsZ)
}
\description{
First sentence. second sentence.
}
\examples{
data(ex.ageBevertonHoltsZ)
head(ex.ageBevertonHoltsZ)


}
\keyword{datasets}
\keyword{length}

