% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.BevertonHoltsZ.R
\docType{data}
\name{ex.BevertonHoltsZ}
\alias{ex.BevertonHoltsZ}
\title{Example dataset of length for use in BevertonHoltsZ}
\format{A list consisting of: 1. a dat}
\usage{
data(ex.BevertonHoltsZ)
}
\description{
First sentence. second sentence.
}
\examples{
data(ex.BevertonHoltsZ)
head(ex.BevertonHoltsZ)


}
\keyword{datasets}
\keyword{length}

