% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.LengthCC.R
\docType{data}
\name{ex.LengthCC}
\alias{ex.LengthCC}
\title{Example dataset of length for use in catch curve}
\format{A list consisting of: 1. a dataframe}
\usage{
data(ex.LengthCC)
}
\description{
First sentence. second sentence.
}
\examples{
data(ex.LengthCC)
head(ex.LengthCC)


}
\keyword{datasets}
\keyword{length}

