% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GillnetSelect.R
\name{GillnetSelect}
\alias{GillnetSelect}
\title{Gillnet selctivity}
\usage{
GillnetSelect(param)
}
\arguments{
\item{param}{A list with following parameters: midlengths of size classes (\code{$midLengths}), number of fish caught with net 1 (\code{$numNet1}), number of fish caught with net 2 (\code{$numNet2}), and the meshsizes of both nets (\code{$msNet1} & \code{$msNet2}).}

\item{numNets}{A number indicating how many nets (with different mesh sizes) were used in the experiment (default: \code{numNets = 2}).}
}
\description{
This function estimates the selecitvity of a gillnet from an experimental catch with two gillnets with different mesh sizes.
}
\details{
This function estimates the fractions retained by each net (SNet1 & SNet2), the
  optimum lengths for each net, the selection factor (SF), and the standard deviation
  of the factor (stand.dev).
  Assumptions of this method are, that (i) the optimum length Lm is proportional to the mesh
  size (Lm = SF * m), (ii) the selection curves are normally distributed with a common
  standard deviation, (iii) the nets have the same fishing power (same dimensions and material).
  Requirements for the experimental set-up are: selection curves corresponding to the two
  mesh sizes have to overlap, and the nets have to be set in the same area, during the
  same time.
}
\examples{
data(data_GillnetSelect)
output <- GillnetSelect(data_GillnetSelect)
plot(output)

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

