% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GillnetSelect.R
\name{GillnetSelect}
\alias{GillnetSelect}
\title{Gillnet selctivity}
\usage{
GillnetSelect(classes, numNet1, numNet2, msNet1, msNet2)
}
\arguments{
\item{classes}{Midpoints of the length class as vector (length frequency data) or ages as vector (age composition data).}

\item{numNet1}{Number of individuals caught using the smaller meshed net}

\item{numNet2}{Number of individuals caught using the larger meshed net}

\item{msNet1}{mesh size of smaller meshed net}

\item{msNet2}{mesh size of larger meshed net}

\item{catch}{Catch as vector, or a matrix with catches of subsequent years if the catch curve with constat time intervals should be applied.}
}
\description{
This function estimates the selecitvity of gillnets nets.
}
\details{
Preconditions: selection curves of two mesh sizes must overlap, nets set in same area, during the same time. To calculate selection factor (SF), L25, L50 and L75 for gillnets. Assumptions: optimum length Lm is proportional to mesh size (Lm = SF * m), two selection curves have the same standard deviation, nets have the same fishing power (same dimensions and material). Assumption that selection curves are normally distributed with a common standard deviation.
}
\examples{
data("ex.GillnetSelect")
with(ex.GillnetSelect,GillnetSelect(midLengths,numCa1,numCa2,
   msNet1 = 8.1,msNet2 = 9.1))
}
\references{
XX
}

