% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{goatfish}
\alias{goatfish}
\title{Yellowstriped goatfish data}
\format{A list consisting of:
\itemize{
  \item \code{midLengths}: mid points of length classes,
  \item \code{catch}: a vector with catches in numbers,
  \item \code{Linf}: infinite length in cm [cm],
  \item \code{K}: growth coefficent per year [1/year].
}}
\source{
Ziegler, B., 1979. Growth and mortality rates of some fishes of
  Manila Bay, Philippines, as estimated from analysis of length-frequencies.
  Thesis. Kiel University, 115 p.
}
\usage{
data(goatfish)
}
\description{
Data of Yellowstriped goatfish (\emph{Upeneus vittatus}) from Manila Bay,
   Philippines. Can be used for the estimation of the instantaneous mortality
   rate (Z) by means of \code{\link{catchCurve}}.
}
\examples{
data(goatfish)
str(goatfish)
summary(goatfish)


}
\keyword{catchCurve}
\keyword{data}
\keyword{dataset}
\keyword{length-frequency}
