% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goatfish.R
\docType{data}
\name{goatfish}
\alias{goatfish}
\title{Data of Yellowstriped goatfish (\emph{Upeneus vittatus}) from Manila Bay, Philippines.}
\format{A list consisting of: 1. a dataframe}
\source{
Ziegler, B., 1979. Growth and mortality rates of some fishes of Manila Bay, Philippines, as estimated from analysis of length-frequencies. Thesis. Kiel University, 115 p.
}
\usage{
data(goatfish)
}
\description{
Data of Yellowstriped goatfish (\emph{Upeneus vittatus}) from Manila Bay,
   Philippines. Can be used for \code{\link{CatchCurve}}.
}
\examples{
data(goatfish)
head(goatfish)


}
\keyword{dataset}
\keyword{length}

