% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_length_age.R
\name{growth_length_age}
\alias{growth_length_age}
\title{Estimation from growth parameter using length-at-age data}
\usage{
growth_length_age(param, method, Linf = NA)
}
\arguments{
\item{param}{a list consisting of following parameters:
\itemize{
  \item \strong{age}: age groups,
  \item \strong{length}: corresponding length groups.
}}

\item{method}{indicating which of following methods should be applied: "GullandHolt",
"FordWalford", "Chapman", or "BertalanffyPlot".}
}
\description{
This function estimates growth parameters from length-at-age data. It
   allows to perform different methods: Gulland and Holt, Ford Walford plot, or
   Chapman's method.
}
\details{
Gulland and Holt plot assumes
infinitestimal delta t (only reasonable approximation of growth parameters if delta t
is small). Ford Walford plot and Chapman assume constant time intervals between ages
(delta t). The Bertalanffy plot is a robust method, however it requires an estimate of Linf. As
long as this estimate is reasonable the resulting estimate of K is reasonable. For
a first estimate of Linf the Powell Wetherall method \link{powell_wetherall} can
be used. Otherwise, the largest fish or the average of the ten largest fish can be
used for a small or large sample, respectively.
}
\examples{
# synthetical length at age data
dat <- list(age = 1:7, length = c(25.7,36,42.9,47.5,50.7,52.8,54.2))
growth_length_age(dat, method = "GullandHolt")

# Bertalaffy plot
dat <- list(age = c(0.64,1.16,1.65,2.1,2.64,3.21),
   length = c(17.3,27.9,35.3,40.2,43.3,45.5))
growth_length_age(dat, method = "BertalanffyPlot", Linf = 50)

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.
}

