% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emperor.R, R/hake.R
\docType{data}
\name{hake}
\alias{hake}
\title{Emperor data}
\format{A dataframe containing: 1. \strong{year} years, 2. \strong{Y} yield,
   3. \strong{CPUE} CPUE or abundance index}
\source{
Dharmendra, D., Sólmundsson, J., 2005. Stock assessment of the offshore Mauritian banks
using dynamic biomass models and analysis of length frequency of the Sky Emperor
(\emph{Lethrinus mahsena}). Fisheries Training Program The United Nations University, 61

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\usage{
data(emperor)

data(hake)
}
\description{
This dataset contains information about sky emperor (\emph{Lethrinus mahsena})
 and its fisheries of offshore Mauritius banks (Nazareth banks). It can be used for several methods,
 e.g. \code{\link{prod_mod_dyn}}.

This dataset contains information about hake (\emph{Merluccius merluccius})
 and its fisheries off Senegal. It can be used for several methods,
 e.g. \code{\link{Predict_ThompsonBell1}}.
}
\examples{

data(emperor)
str(emperor)



data(hake)
str(hake)


}
\keyword{data}
\keyword{dataset}
\keyword{emperor}
\keyword{hake}

