% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LengthConCatchCurve.R
\name{LengthConCatchCurve}
\alias{LengthConCatchCurve}
\title{Length converted catch curve}
\usage{
LengthConCatchCurve(midLengths, catch, Linf, K, t0 = 0, catchCorFac = NA)
}
\arguments{
\item{midLengths}{The midpoints of the size class intervals as vector.}

\item{catch}{The catch per sampling time as matrix or the total catch as vector.}

\item{Linf}{The infinite length for investigated species in cm [cm].}

\item{K}{The growth coefficient for investigated species per year [1/year].}

\item{t0}{The theoretical time zero, at which individuals of this species hatch.}

\item{catchCorFac}{optional: Correction factor for catch, in case catch does not reflect catch of a whole year or it reflects only part of fishing ground.}
}
\description{
This is a function to calculate the total mortality (Z) from length composition data via the length converted catch curve.
}
\details{
Catch curve..., results (Dataframe, Total mortality, Plot) are saved in the list "results.CC".
}
\examples{
\donttest{
 data("ex.LengthCC")
 output <- LengthConCatchCurve(ex.LengthCC[,1], ex.LengthCC[,2], Linf = 23.1, K = 0.59)
 output
 }
}

