% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{lfqCohort}
\alias{lfqCohort}
\title{Assign cohort, relative ages and birth date to an lfq dataset}
\usage{
lfqCohort(lfq, n.cohort.per.yr = 1, agemax = NULL, calc_dt = FALSE)
}
\arguments{
\item{lfq}{an lfq object with fitted VBGF parameters in the \code{lfq$par} slot.}

\item{n.cohort.per.yr}{number of cohorts per year for slicing (Default:
 \code{n.cohort.per.year  = 1}).
See details section for more information on slicing.
This argument should be set to a higher values in cases of multiple cohorts per year.}

\item{agemax}{maximum age of the stock, which is used to define the extent of growth
curves via a call to \code{\link{lfqFitCurves}}. When not supplied
(Default: \code{agemax = NULL}), the value is estimated based on the time rquired
to achieve a length of 95\% of Linf.}

\item{calc_dt}{logical. Should time required to grow through bin be calculated.
(Default: \code{calc_dt = FALSE}).}
}
\value{
a list of class `lfq` containing the additional slots for relative age
(\code{rel.age}), cohort number (\code{cohort}), birthday (\code{bday}), and
(optionally) time required to growth through bin (\code{dt}).
}
\description{
The \code{lfqCohort} function will assign cohort number (1 = oldest),
relative age (since length equal zero), and birth date (at length equal zero) to
an \code{lfq} object given a set of VBGF parameters (\code{lfq$par}).

The method is can be used as a precursor to cohort-based analyses
(e.g. catch curve, VPA), or as means of identifying recruitment patterns.
Relative ages consider seasonalized VBGF (when applicable), thus aiding the
correct assignment of ages to lfq data (see Pauly 1990).
}
\details{
The method involves "slicing" the lfq data based on the VBGF parameters, and
follows the general approach described by Pauly (1990), and demonstrated by
Pauly et al. (1995) using the "GOTCH.A" program. By default, a single slice per year
(\code{n.cohort.per.yr = 1}) is used, which sets slice boundaries of +/- 0.5 yr around the
\code{ta} parameter. For each lfq bin, cohort association is then determined by
its inclusion within a given slice. Bins of length greater than Linf, are aggregated with
the bin that includes Linf. In the case of multiple cohorts per year, a larger
\code{n.cohort.per.yr} is needed in order to resolve intra-annual cohorts.
}
\examples{
# Load data and add VGBF parameters
data(synLFQ4)
synLFQ4$par <- list(Linf = 80, K = 0.5, ta = 0.25, C = 0.75, ts = 0.5)

# visualize cohorts with lfqCohort
synLFQ4 <- lfqCohort(synLFQ4, n.cohort.per.yr = 1)
plot(synLFQ4, Fname = "catch",
  ylim = c(0, max(synLFQ4$midLengths)), image.col=NA)
pal <- colorRampPalette(c(4,5,7,2))
with(synLFQ4, image(x = dates, y = midLengths, z = t(cohort),
  col = adjustcolor(pal(max(cohort, na.rm = TRUE)), 0.75),
  add=TRUE
))

# visualize relative age with lfqCohort
plot(synLFQ4, Fname = "catch",
  ylim = c(0, max(synLFQ4$midLengths)), image.col=NA)
pal <- colorRampPalette(c(4,5,7,2))
with(synLFQ4, image(x = dates, y = midLengths, z = t(rel.age),
  col = adjustcolor(pal(100), 0.75),
  add=TRUE
))


}
\references{
Pauly, D., Moreau, J., & Abad, N. (1995). Comparison of age-structured
and length-converted catch curves of brown trout Salmo trutta in two
French rivers. Fisheries Research, 22(3-4), 197-204.
https://doi.org/10.1016/0165-7836(94)00323-O

Pauly, Daniel. (1990). Length-converted catch curves and the seasonal
growth of fishes. Fishbyte, 8(3), 33-38.
}
