% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_Pauly.R
\name{M_Pauly}
\alias{M_Pauly}
\title{Pauly's empirical formula for the estimation of M}
\usage{
M_Pauly(Linf = NA, Winf = NA, K, temp, schooling = FALSE)
}
\arguments{
\item{Linf}{infinite total length (TL) for investigated species in cm [cm].}

\item{Winf}{infinite weight for investigated species in g [g].}

\item{K}{growth coefficient for investigated species per year [1/year].}

\item{temp}{average annual temperature at the surface in degrees centigrade.}

\item{schooling}{logical; if TRUE it is accounted for the schooling behaviour of
the species. Default is FALSE.}
}
\value{
A numeric value representing M per year [1/year].
}
\description{
Function to calculate the instantaneous natural mortality rate (M)
     according to Pauly's empirical formula.
}
\details{
This function can either be applied to the infinite length or to the
   infinite weight. If accounting for schooling behaviour M is multiplied by 0.8 according to
   Pauly (1983).
}
\examples{
M_Pauly(Linf = 80, K = 0.5, temp = 25)

}
\references{
Pauly, D., 1980. On the interrelationships between natural mortality, growth parameters,
and mean environmental temperature in 175 fish stocks. \emph{J.Cons.CIEM}, 39(2):175-192

Pauly, D., 1983. Some simple methods for the assessment of tropical fish stocks.
\emph{FAO Fish.Tech.Pap.}, (234): 52p. Issued also in French and Spanish

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.
}
\keyword{M}
\keyword{function}
\keyword{mortality}

