% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_Pauly.R
\name{M_Pauly}
\alias{M_Pauly}
\title{Pauly's empirical formula for natural mortality (M)}
\usage{
M_Pauly(Linf, K, temp, schooling = FALSE)
}
\arguments{
\item{Linf}{Infinite length for investigated species in cm [cm].}

\item{K}{Growth coefficient for investigated species per year [1/year].}

\item{schooling}{Logical parameter accounting for schooling behaviour, according to Pauly (1983). Default is FALSE.}

\item{Temp}{Average annual temperature at the surface in degrees centigrade.}
}
\value{
A numeric value representing the natural mortality (M).
}
\description{
Pauly's empirical formula to calculate natural mortality (M).
}
\details{
Pauly's empirical formula to calculate natural mortality (M).
   If accounting for schooling behaviour M is multiplied by 0.8 according to Pauly (1983).
}
\examples{
M_Pauly(Linf = 80, K = 0.5, temp = 25)

}
\references{
Pauly, D., 1980. On the interrelationships between natural mortality, growth parameters,
and mean environmental temperature in 175 fish stocks. \emph{J.Cons.CIEM}, 39(2):175-192

Pauly, D., 1983. Some simple methods for the assessment of tropical fish stocks.
\emph{FAO Fish.Tech.Pap.}, (234): 52p. Issued also in French and Spanish

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\keyword{M}
\keyword{function}
\keyword{mortality}

