% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_RikhterEfanov.R
\name{M_RikhterEfanov}
\alias{M_RikhterEfanov}
\title{Rikhter and Efanov's formula}
\usage{
M_RikhterEfanov(tm50)
}
\arguments{
\item{tm50}{age when 50\% of the population is mature [year] ("age of massive maturation")}
}
\value{
A numeric value representing the natural mortality (M).
}
\description{
Rikhter and Efanov's formula to calculate natural mortality (M).
}
\details{
Rikhter and Efanov's formula to calculate natural mortality (M).
}
\examples{
M_RikhterEfanov(tm50 = 1)

}
\references{
Rikhter, V.A., and V.N. Efanov, 1976. On one of the approaches to estimation of natural
mortality of fish populations. \emph{ICNAF Res.Doc.}, 76/VI/8: 12p.

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\keyword{M}
\keyword{function}
\keyword{mortality}

