% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.catchCurve.R
\name{plot.catchCurve}
\alias{plot.catchCurve}
\title{Plotting catch curve}
\usage{
\method{plot}{catchCurve}(x, plot.selec = FALSE, ...)
}
\arguments{
\item{x}{A list of the class \code{"catchCurve"} containing the results of the
catchCurve model.}

\item{plot.selec}{logical; if TRUE the regression line is plotted for not fully
exploited length groups and the probability of capture is plotted}

\item{...}{normal parameters from plot function}
}
\description{
This function plots the results from the \link{catchCurve} model.
}
\details{
A function to plot the results of the catchCurve model.
}
\examples{
\donttest{
#_______________________________________________
# Variable paramter system (with catch vector)
# based on length frequency data
# load data
data(goatfish)

# run model
catchCurve(goatfish,calc_ogive=TRUE)

# based on age composition data
# load data
data(whiting)

# run model
catchCurve(whiting, catch_column = 1)

#_______________________________________________
# Catch Curve with estimation of selection ogive
# load data
data(synLFQ3)

# run model
output <- catchCurve(synLFQ3, calc_ogive = TRUE)

# plot results
plot(output)
}

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

