% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CatchCurve.R
\name{plot.CatchCurve}
\alias{plot.CatchCurve}
\title{Catchcurve plot}
\usage{
\method{plot}{CatchCurve}(x, ...)
}
\arguments{
\item{x}{A list of the class \code{"CatchCurve"} containing the results of the CatchCurve model.}

\item{...}{normal parameters from plot function}
}
\description{
This function plots the results from the \link{CatchCurve} model.
}
\details{
A function to plot the results of the CatchCurve model.
}
\examples{
\donttest{
data(whiting)
output <- CatchCurve(whiting, catch_column = 1)
plot(output)
}

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

