% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GillnetSelect.R
\name{plot.GillnetSelect}
\alias{plot.GillnetSelect}
\title{Gillnet selectivity plot}
\usage{
\method{plot}{GillnetSelect}(param, ...)
}
\arguments{
\item{param}{List with parameters: midlengths of size classes, number in net, number in codend and meshsize of codend}
}
\description{
This function estimates the selecitvity of trawl nets.
}
\details{
To calculate selection factor (SF), L25, L50 and L75 for trawl nets /fisheries.
}
\examples{
data(data_GillnetSelect)
output <- GillnetSelect(data_GillnetSelect)
plot(output)

}
\references{
xx
}

