% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GillnetSelect.R
\name{plot.GillnetSelect}
\alias{plot.GillnetSelect}
\title{Gillnet selectivity plot}
\usage{
\method{plot}{GillnetSelect}(x, ...)
}
\arguments{
\item{x}{A list of the class "GillnetSelect" containing the results of the gillnet selectivity function.}

\item{...}{normal parameters from plot function}
}
\description{
This function plots the selectivity estimates of the
  function \code{\link{GillnetSelect}}. Second sentence.
}
\details{
A function to plot the results of the gillnet selectivity estimation.
}
\examples{

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

