% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MillarsGillnet.R
\name{plot.MillarsGillnet}
\alias{plot.MillarsGillnet}
\title{Millar's gillnet selectivity plot}
\usage{
\method{plot}{MillarsGillnet}(x, ret_dev = c(T, T), cex = 1,
  title = "Deviance residuals", ...)
}
\arguments{
\item{x}{A list of the class "GillnetSelect" containing the results of the gillnet selectivity function.}

\item{ret_dev}{draw both plots? Default = c(T,T)}

\item{cex}{thickness of points Default = 1}

\item{title}{Title of plot Default = "Deviance residuals"}

\item{...}{normal parameter options from plot function}
}
\description{
This function plots the selectivity estimates of the function
   \code{\link{GillnetSelect}}. First sentence. second sentence.
}
\details{
A function to plot the results of the gillnet selectivity estimation.
}
\examples{
# load data
data(data_GillnetSelect3)

# run model
output <- MillarsGillnetSelect(data_GillnetSelect3, model = "normal_fixed",
   plotlens = NULL, rel = NULL)

# investigate results
output

#plot results
plot(output)

}
\references{
Millar, R. B., Holst, R., 1997. Estimation of gillnet and hook selectivity
 using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.
}

