% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predict_mod.R
\name{plot.predict_mod}
\alias{plot.predict_mod}
\title{Plotting prediction models yield per recruit and Thompson & Bell}
\usage{
\method{plot}{predict_mod}(x, type = "ypr", xaxis1 = "FM", yaxis1 = "Y_R",
  yaxis2 = "B_R", ...)
}
\arguments{
\item{x}{a object of the class 'predict_mod'}

\item{type}{default = "ypr" or "Isopleth" for isopleth plot}

\item{xaxis1}{default = "FM" which x axis should be plotted? FM or E?}

\item{yaxis1}{default = "Y_R" for ypr plot, which first y axis should be plotted? Y_R or Y_R.rel}

\item{yaxis2}{default = "B_R" for ypr plot, which second y axis should be plotted? B_R or B_R.percent}

\item{...}{optional parameters of plot function}
}
\description{
This function plots objects of the class "predict_mod", which are results
   of the function \code{\link{predict_mod}}.
}
\examples{
# YPR
# age structured data
# Nemipterus marginatus
# threadfin <- list(Winf = 286,K = 0.37, t0 = -0.2, M = 1.1, tr = 0.4)

# run model
# predict_mod(threadfin, FM_change = seq(0,6,0.1),
#    Lc_tc_change = seq(0.2,1,0.2), type = 'ypr')  #where it is maximal  = MSY

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.

# @export
}

