% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.prod_mod_ts.R
\name{plot.prod_mod_ts}
\alias{plot.prod_mod_ts}
\title{Plotting time series production models}
\usage{
\method{plot}{prod_mod_ts}(x, ...)
}
\arguments{
\item{x}{a object of the class 'prod_mod_dyn',}

\item{...}{optional parameters of plot function}
}
\description{
This function plots objects of the class "prod_mod_ts".
}
\examples{
# load data
data(emperor)

# run model
#output <-  prod_mod_ts(data = emperor, method  = "Fox")

# plot output
#plot(output)


}
\references{
Dharmendra, D., Solmundsson, J., 2005. Stock assessment of the offshore Mauritian banks
using dynamic biomass models and analysis of length frequency of the Sky Emperor
(\emph{Lethrinus mahsena}). Fisheries Training Program The United Nations University, 61

#@export
}

