% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.select_Millar.R
\name{plot.select_Millar}
\alias{plot.select_Millar}
\title{Millar's selectivity plot}
\usage{
\method{plot}{select_Millar}(x, plotlens = NULL, standardize = TRUE, ...)
}
\arguments{
\item{x}{A list of the class "GillnetSelect" containing the results of the gillnet selectivity function.}

\item{...}{normal parameter options from plot function}

\item{ret_dev}{draw both plots? Default = c(T,T)}

\item{cex}{thickness of points Default = 1}

\item{title}{Title of plot Default = "Deviance residuals"}
}
\description{
This function plots the selectivity estimates of the function
   \code{\link{GillnetSelect}}. First sentence. second sentence.
}
\details{
A function to plot the results of the gillnet selectivity estimation.
}
\examples{
# Gillnet
# load data
data(gillnet)

# run model
output <- select_Millar(gillnet, x0 = c(60,4), rel.power = rep(1,8),
   rtype = "norm.loc")

# investigate results
output

# plot results
plot(output,plotlens=seq(40,90,0.1))

}
\references{
Millar, R. B., Holst, R., 1997. Estimation of gillnet and hook selectivity
 using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.
}

