% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TrawlSelect.R
\name{plot.TrawlSelect}
\alias{plot.TrawlSelect}
\title{Trawl net selectivity plot}
\usage{
\method{plot}{TrawlSelect}(x, ...)
}
\arguments{
\item{x}{Results of the trawl select function in the class \code{"TrawlSelect"}.}

\item{...}{parameter of plot function}
}
\description{
This function estimates the selecitvity of trawl nets.
   First sentence. second sentence.
}
\details{
To calculate selection factor (SF), L25, L50 and L75 for trawl nets /fisheries.
}
\examples{
# load data
data(bream)

# run model
output <- select(bream)

# plot results
#plot(output)

}
\references{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}

