% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mod.R
\name{predict_mod}
\alias{predict_mod}
\title{Prediction models: YPR and Thompson & Bell model}
\usage{
predict_mod(param, FM_change = NA, Lc_tc_change = NULL, type, s_list,
  stock_size_1 = NA, unit.time = "year", curr.E = NA, curr.Lc_tc = NA,
  plus.group = NA)
}
\arguments{
\item{param}{a list consisting of following parameters (not all are required):
\itemize{
  \item \strong{Linf} or \strong{Winf}: infinite length for investigated species in cm [cm],
  \item \strong{K}: growth coefficent for investigated species per year [1/year],
  \item \strong{t0}: theoretical time zero, at which individuals of this species hatch,
  \item \strong{M}: natural mortality,
  \item \strong{Z}: total mortality,
  \item \strong{FM}: fishing mortality,
  \item \strong{a}: ,
  \item \strong{b}: ,
  \item \strong{Lr} or \strong{tr}: length or age of recuritment;}
additional list objects in for the Thompson and Bell model:
 \itemize{
  \item \strong{age} or \strong{midLengths}: midpoints of the length class as vector (length-frequency
  data) or ages as vector (age composition data),
  \item \strong{meanWeight}: vector with mean weight per length group or age class,
  \item \strong{meanValue}: vector with mean value per length group or age class,
}}

\item{FM_change}{vector with ascending fishing mortalities}

\item{Lc_tc_change}{vector with ascending lengths or ages at first capture (Lc/tc)}

\item{type}{indicating which model should be applied: \code{"ypr"} for Beverton and Holt's
yield per recruit model and \code{"ThompBell"} for the Thompson and Bell model}

\item{s_list}{list with selectivity parameters}

\item{stock_size_1}{stock size of smallest size class, if NA values are calculated relative to a stock size of 1000 individuals}

\item{unit.time}{in which time unit the data is provided? "month" or "year"}

\item{curr.E}{current exploitation rate}

\item{curr.Lc_tc}{current Lc (length at first capture)}

\item{plus.group}{if a value is provided, a plus group is created comprising this size class and all above}
}
\value{
A list with the input parameters and dependent on the model type following
   list objects:
\itemize{
  \item \code{type = 'ypr'}
  \itemize{
     \item \strong{FM}: fishing mortalities,
     \item \strong{Lc} or \strong{tc}: lengths or ages at first capture,
     \item \strong{list_Lc_runs} or \strong{list_tc_runs}: a list with the dataframes for each Lc or tc value:
     \itemize{
     \item FM_change
     \item Y_R
     \item Y_R.rel
     \item B_R
     \item B_R.percent
     \item Ty
     \item Wy
     }
  }
  \item \code{type = 'ThomBell'}
  \itemize{
     \item \strong{dt}: delta t,
     \item \strong{N}: population number,
     \item \strong{dead}: deaths due to natural reasons,
     \item \strong{C}: catch,
     \item \strong{Y}: yield,
     \item \strong{B}: biomass,
     \item \strong{V}: value,
     \item \strong{totals}: summed up values (total catch, total yield, total value, average biomass),
     \item \strong{tot.C}: total catches for different x factors,
     \item \strong{tot.Y}: total yield values for different x factors,
     \item \strong{tot.V}: total values for different x factors,
     \item \strong{meanB}: average biomasses for different x factors,
     \item \strong{Xfact}: fishing mortality changes;
  }
  \item \code{type = 'ThomBell'} and \code{Lc_tc_change} provided
  \itemize{
     \item \strong{FM_change}: fishing mortality changes,
     \item \strong{Lc_tc_change}: changes in length or age at first capture,
     \item \strong{Lt}: lengths at age,
     \item \strong{sel}: probability of capture,
     \item \strong{mat_FM_Lc_com.C}: catch matrix for all fishing mortality and Lc/tc combinations,
     \item \strong{mat_FM_Lc_com.Y}: yield matrix for all fishing mortality and Lc/tc combinations,
     \item \strong{mat_FM_Lc_com.V}: value matrix for all fishing mortality and Lc/tc combinations,
     \item \strong{mat_FM_Lc_com.B}: biomass matrix for all fishing mortality and Lc/tc combinations;
  }

}
}
\description{
This is a function predicting catch, yield, biomass and economic values for different
   fishing mortality scenarions (in combination with gear changes). It allows to choose if the
   calculation should be based upon Beverton and Holt's yield per recruit model or the
   more sophisticated Thompson and Bell model.
}
\details{
better to treat last group always as a plus group.....
   The Thompson and Bell model incorporates an iteration step simulating the stock by means
   of the \code{\link{stock_sim}} function. In case changes in gear characteristics -
   here measured in terms of Lc or tc, the length or age at first capture, respectively -
   should be explored, a list with selectivity information about the gear has to be provided and
   the prediction models make use of the selectivity \code{\link{select_ogive}} function.
}
\examples{
#______________________________________
#       Yiel Per Recruit (YPR)
#______________________________________
# age structured data
# Nemipterus marginatus
threadfin <- list(Winf = 286,K = 0.37, t0 = -0.2, M = 1.1, tr = 0.4)

# run model
predict_mod(threadfin, FM_change = seq(0,6,0.1),
   Lc_tc_change = seq(0.2,1,0.2), type = 'ypr')  #where it is maximal  = MSY

# Leiognathus spendens (Pauly 1980)
ponyfish <- list(Winf = 64, K = 1, t0 = -0.2, M = 1.8, tr = 0.2)

# run model
predict_mod(ponyfish, Lc_tc_change = c(0.2,0.3,1.0), type = 'ypr')

#______________________________________
# length structured data
# Xiphias gladius (Berkeley and Houde 1980)
swordfish <- list(Linf = 309, K = 0.0949, M = 0.18, a=0.0003, b=3, Lr = 90)  ## T_Lr , a, b ??? assumed

# run model
predict_mod(swordfish, Lc_tc_change = c(100,118,150,180), type = 'ypr')

####test: E <- seq(0,0.9,0.1) FM <- E * M / (1 - E)


#______________________________________
#      Thompson and Bell model
#______________________________________
# with age structured data
# load data
data(shrimps)

# run model
predict_mod(shrimps, FM_change = seq(0.1,3,0.1), type = 'ThompBell')

# create list with selectivity information
select.list <- list(selecType = 'knife_edge',  #or 'gillnet' or 'trawl_ogive'
   Lc = 34, tc = 5, selecDist = 'lognormal',    #or 'normal_fixed'
   mesh_size = 8.1, mesh_size1 = 9.1, select_p1 = 21.1, select_p2 = 23.8)

# add additional parameters to data list
shrimps <- c(shrimps,  list(Linf = 50, K = 0.3, t0 = 0.01))

# run model
predict_mod(shrimps,FM_change = seq(0,3,0.2), Lc_tc_change = seq(24,44,2),
   type = 'ThompBell', s_list = select.list, unit.time = 'month')

#______________________________________
# with length structured data
# load data
data(hake)

# run model
predict_mod(hake,FM_change = seq(0.1,3,0.1), type = 'ThompBell')


}
\references{
example 1 : Kuwait (Garcia and van Zalinge 1982)

Beverton and Holt 1966

Boerema, L.K., and J.A. Gulland, 1973. Stock assessment of the Peruvian anchovy
(Engraulis ringens) and management of the fishery. Journal of the Fisheries Board of
Canada, 30(12):2226-2235

Millar, R.B., and R. Holst, 1997. Estimation of gillnet and hook selectivity using
log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3):471-477

Sparre, P., and Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\keyword{function}
\keyword{prediction}

