% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_ThompsonBell2.R
\name{Predict_ThompsonBell2}
\alias{Predict_ThompsonBell2}
\title{Thompson and Bell prediction model}
\usage{
Predict_ThompsonBell2(param, FM_change)
}
\arguments{
\item{param}{A list containing all information}

\item{FM_change}{Vector with ascending Fishing mortalities}
}
\description{
This is a function to calculate the total mortality (Z) from
   length composition data via the length converted catch curve or from age at
   length data with catch curve.
}
\details{
better to treat last group always as a plus group..... For variable parameter system vectors are reuqired for constant parameter systems matrices or data.frames have to be inserted. or vectors The length converted linearised catch curve is used to calculate the total mortality (Z). This function includes a so called locator function, which asks you to choose points from a graph manually. Based on these points the regression line is calculated.
}
\examples{

# yield per recruit with age structured data
# load data
data(shrimps)

# run model
output <- Predict_ThompsonBell2(shrimps,seq(0.1,3,0.1))

# investigate results
output

# yield per recruit with length structured data
# load data
data(hake)

# run model
Predict_ThompsonBell2(hake,seq(0.1,3,0.1))

}
\references{
example 1 : Kuwait (Garcia and van zalinge 1982)
  Millar, R. B., & Holst, R. (1997). Estimation of gillnet and hook selectivity using log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3), 471-477.
}

