% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_mod.R
\name{prod_mod}
\alias{prod_mod}
\title{Production models - Schaefer and Fox models}
\usage{
prod_mod(data)
}
\arguments{
\item{data}{a dataframe consisting of:
\itemize{
  \item \strong{year} years,
  \item \strong{Y} catch in weight per year, and
  \item \strong{f} fishing effort per year, or
  \item \strong{CPUE} catch per unit of effort per year.
}}
}
\description{
Surplus production models (or production models) are holistic models, which
   can be used to estimate the maximum sustainable yield, biomass and effort for a
   certain population. This function applies two models simultaneously: the Schaefer
   and the Fox model.
}
\details{
Production models can be applied if sufficient data are available: effort
   and yield parameters have to be expended over a certain number of years. Furthermore,
   the fishing effort must have undergone substantial changes over the period covered
   (Sparre and Venema, 1998). Either the Catch per unit of effort (CPUE) is inserted
   into the model directly (by a column \strong{CPUE}) or the CPUE is calculated from
   the catch and effort, then these two vectors should have required units.
}
\examples{
# load data
data(trawl_fishery_Java)

# run model
prod_mod(data = trawl_fishery_Java)

}
\references{
Fox, W. W. Jr., 1970. An exponential surplus-yield model for optimizing exploited fish
populations. \emph{Trans.Am.Fish.Soc.}, 99:80-88

Graham, M., 1935. Modern theory of exploiting a fishery and application to North Sea
trawling. \emph{J.Cons.CIEM}, 10(3):264-274

Schaefer, M., 1954. Some aspects of the dynamics of populations important to the
management of the commercial marine fisheries. \emph{Bull.I-ATTC/Bol. CIAT}, 1(2):27-56

Schaefer, M., 1957. A study of the dynamics of the fishery for yellowfin tuna of the
eastern tropical Pacific Ocean [in English and Spanish]. \emph{Ibid.}, 2(6): 245-285

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\keyword{MSY}
\keyword{biomass}
\keyword{function}
\keyword{stock}

