% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_mod_dyn.R
\name{prod_mod_dyn}
\alias{prod_mod_dyn}
\title{Non-equilibrium dynamic biomass models}
\usage{
prod_mod_dyn(data)
}
\arguments{
\item{data}{a dataframe of parameters
\itemize{
  \item \strong{year} years,
  \item \strong{yield} catch in weight of fishery per year, and
  \item \strong{effort} fishing effort per year, or
  \item \strong{CPUE} catch per unit of effort per year.
}}
}
\description{
Non-equilibrium dynamic biomass models
}
\details{
Either the Catch per unit of effort (CPUE) is inserted
   into the model directly (by a column \strong{CPUE}) or the CPUE is calculated from
   the catch and effort, then these two vectors should have required units.
}
\examples{
# load data
data(emperor)

# run model
prod_mod_dyn(data = emperor)

}
\references{
Dharmendra, D., Sólmundsson, J., 2005. Stock assessment of the offshore Mauritian banks
using dynamic biomass models and analysis of length frequency of the Sky Emperor
(\emph{Lethrinus mahsena}). Fisheries Training Program The United Nations University, 61

Hilborn, R. and Walters, C., 1992. Quantitative Fisheries Stock Assessment: Choice,
Dynamics and Uncertainty. Chapman and Hall, New York
}
\keyword{MSY}
\keyword{biomass}
\keyword{function}
\keyword{stock}

