% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment}
\alias{recruitment}
\title{Calculate recruitment pattern of lfq object using lfqCohort approach}
\usage{
recruitment(lfq, n.cohort.per.yr = 36, agemax = NULL, plot = TRUE,
  use.rcounts = FALSE, ...)
}
\arguments{
\item{lfq}{a length-frequency object (i.e. class \code{lfq}) that contains
a \code{lfq$par} slot with \code{\link{VBGF}} parameters (usually fit
via \code{\link{ELEFAN}}, \code{\link{ELEFAN_GA}}, or \code{\link{ELEFAN_SA}}
functions).}

\item{n.cohort.per.yr}{number of cohorts per year for slicing
(Default: \code{n.cohort.per.yr = 36}). Higher values may increase the resolution of the
recruitment estimate,  although at the expense of computation time.}

\item{agemax}{maximum age of the stock, which is used to define the extent of growth
curves via a call to \code{\link{lfqFitCurves}}. When not supplied
(Default: \code{agemax = NULL}), the value is estimated based on the time rquired
to achieve a length of 95\% of Linf.}

\item{plot}{logical. Should monthly recruitment pattern [\%] by plotted
as a \code{\link[graphics]{barplot}} (default: \code{plot = TRUE}).}

\item{use.rcounts}{logical. Should restructured counts
(\code{lfq$rcounts}), rather than raw catch counts (\code{lfq$catch})
be used to derive recruitment pattern. If true, only bins with positive restructures
frequencies are considered, and those values are multiplied by 100
in order to provide appropriate scaling for density approxmations via
\code{\link[graphics]{hist}}.}

\item{...}{further arguments passed to
\code{\link[graphics]{barplot}} when \code{plot = TRUE}.}
}
\value{
object of "histogram" class containing
monthly recruitment pattern
}
\description{
The \code{recruitment2} function calculates the
recruitment pattern (i.e. monthly) of an lfq object given
corresponding \code{\link{VBGF}} parameters.
Time at recruitment is calculated via a call to
\code{\link{lfqCohort}}, which slices the lfq data, and assigns
bins to a given cohort, with associated relative ages and brithdays.

This function does not represent a robust statistical method as it
assumes the same growth parameters for all individual counts
in the \code{lfq$catch} bins.
Nevertheless, the results should provide general information about
the recruitment pattern, e.g. relative recruitment strength by month as
weighted by the catch frequencies (\code{lfq$catch}).
Using the argument \code{use.rcounts = TRUE}, the recruitment pattern
is weighted by positive restructured frequencies only.
}
\details{
The method uses the \code{lfqCohort} function to "slice" lfq data
into distinct cohorts.
By setting the number of cohorts per year to a high value
(e.g. \code{n.cohort.per.yr = 36}), specific bins get associated with a precise
\code{ta} value, while other VBGF parameters are held constant.
Given that the size of recruitment can be subjective,
recruitment patterns should be interpreted as relative months since the
assumption of length zero as a recruitment size is unrealistic, as are the
associated precise recruitment times. Nevertheless, the overall pattern
(e.g. one mode vs two), should inform the resolution at which to define
cohorts (i.e. via \code{\link{lfqCohort}}).
}
\examples{

# using catch frequencies
data(synLFQ4)
synLFQ4$par <- list(Linf = 80, K = 0.5, ta = 0.25, C = 0.75, ts = 0.5)
res <- recruitment(synLFQ4)

# using restructured frequencies
synLFQ4 <- lfqRestructure(synLFQ4, MA = 11)
res <- recruitment(lfq = synLFQ4, use.rcounts = TRUE)

# use of results in external plotting (via histogram())
plot(res, col = 5, freq = FALSE,
  main = "Recruitment pattern for synLFQ4",
  xlab = "Relative month"
)

# presentation as percentages
barplot(res$counts/sum(res$counts)*100, names.arg = month.abb,
  las = 2, ylab = "Recruitment [\%]",
  main = "Recruitment pattern for synLFQ4"
)


}
