% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depletion_mod.R, R/growth_tagging.R, R/recruitment.R
\name{recruitment}
\alias{recruitment}
\title{Depletion model}
\usage{
recruitment(param, tsample, catch_column = NA, plot = FALSE)

recruitment(param, tsample, catch_column = NA, plot = FALSE)

recruitment(param, tsample, catch_column = NA, plot = FALSE)
}
\arguments{
\item{param}{a list}

\item{tsample}{sampling time corresponding to time when catch was sampled as
fraction of year
(e.g. 0.25 for 1st of April). If catch was sampled several times (catch as matrix)
vector has to be provided with sampling times in equal order.}

\item{catch_column}{numeric; if catch in param is a matrix, this number indicates
the column of the catch matrix which should be used for the analysis.}

\item{plot}{logical; indicating whether a plot should be printed}

\item{param}{a list}

\item{param}{a list consisting of following parameters:
\itemize{
  \item \code{midLengths}: midpoints of the length classes (length-frequency
  data),
  \item \code{Linf}: infinite length for investigated species in cm [cm],
  \item \code{K}: growth coefficent for investigated species per year [1/year],
  \item \code{t0}: theoretical time zero, at which individuals of this species hatch,
  \item \code{D}: optional; for generalised vBGF
  \item \code{C}: optional; intensity of the (sinusoid) growth oscillations of
  the model,
  \item \code{ts}: optional; onset of the first oscillation with regard to t = 0,
  \item \code{catch}: catches, vector or matrix with catches of subsequent sampling
  times
}}
}
\value{
A list with the input parameters and following list objects:
\itemize{
  \item \strong{ti}: actual age,
  \item \strong{tS_frac}: age at which the lenght was zero expressed as fraction
  of the year,
  \item \strong{correspond_month}: corresponding month,
  \item \strong{N_months_all}: dataframe with months and the number of recruits
  per month,
  \item \strong{N_months_per}: precentage of number of recruits per month.
}
}
\description{
applies Leslie Davis model

This function estimates growth parameters from tagging data

This function estimates recrutiment patterns from length-frequency data-
}
\details{
should be possible to extract recruitment information from length frequency data (Pauly 1982),
but two main assumptions seldomly met:
 1. all fish in a data set grow as described by a single set of growth parameters
2. one month out of twelve always has zero recruitment (not met, because in temperate waters probable that more than one month zero recruitment and in tropics maybe always recruitment)

Recuritment patterns obtained by backward projection onto the length axis of a set of
length frequency data. The steps are:
 1.
projection onto the time axis of the frequencies after they have been divided by delta t,
the time needed to grow through the length class. (Alternatively, FiSAT allows,
following Moreau and Cuende (1991) the use of restructured frequencies (as used for
untransformed L/F data are used, thus partly compensating for the "spread"
illustrated in Fig. 5.19.);

2.
summation for each month (and irrespective of year) of the adjusted frequencies
projected onto each month;
3.
subtraction, from each monthly sum, of the lowest monthly sum to obtain a zero
value where apparent recruitment is lowest;
4.
output of monthly relative recruitment as a percentage of annual recruitment.

results pertain to months of the year only if precise etimate for t0 is given. (time
lag between peak spawning and recruitment can be used ot obtain value for t0)

should be possible to extract recruitment information from length frequency data (Pauly 1982),
but two main assumptions seldomly met:
 1. all fish in a data set grow as described by a single set of growth parameters
2. one month out of twelve always has zero recruitment (not met, because in temperate waters probable that more than one month zero recruitment and in tropics maybe always recruitment)

Recuritment patterns obtained by backward projection onto the length axis of a set of
length frequency data. The steps are:
 1.
projection onto the time axis of the frequencies after they have been divided by delta t,
the time needed to grow through the length class. (Alternatively, FiSAT allows,
following Moreau and Cuende (1991) the use of restructured frequencies (as used for
untransformed L/F data are used, thus partly compensating for the "spread"
illustrated in Fig. 5.19.);

2.
summation for each month (and irrespective of year) of the adjusted frequencies
projected onto each month;
3.
subtraction, from each monthly sum, of the lowest monthly sum to obtain a zero
value where apparent recruitment is lowest;
4.
output of monthly relative recruitment as a percentage of annual recruitment.

results pertain to months of the year only if precise etimate for t0 is given. (time
lag between peak spawning and recruitment can be used ot obtain value for t0)

This function allows to extract information about the recruitment patterns of a
stock by backward projection onto the length axis of a set of length frequency data
using the (special, generalised or seasonalised)
von Bertallanfy growth curve (vBGF, Pauly 1982). The method assumes that (i) all fish in a
data set grow as described by a single set of growth parameters and (ii) one month out
of twelve always has zero recruitment. The second assumption is probably not met as
in temperate waters more than one month has zero recruitment and in the tropics
there might always be recruitment.
If t0 is not provided, only a realtive recruitment pattern can be estimated. It is
therefore not possible to assign months to the results. However, an estimate of t0
can be obtained by the time
lag between peak spawning and recruitment. Several length-frequency data sets
can be used to estimate the recrutiment pattern by providing catch as a matrix and
setting catch_column to NA (default). Then the fraction per time is calculated for
each size class in each sample an then pooled together. For the generalised vBGF D is
required, for the seasonalised vBGF C, ts and D.
}
\examples{
data(trout)

trout
Linf = 16
K = 0.4

data(trout)

trout
Linf = 16
K = 0.4

# one sample
dat <- list(midLengths = seq(2,98,4),
               catch = c(0.6,17.6,93,83.2,12.6,0.3,0,0,0,1,17.1,51.4,
               26.1,2.2,0.2,4.5,21.6,17.6,3.7,8.7,10.6,6.3,5.6,2.9,0.8),
               Linf = 100,
               K = 0.5,
               t0 = 0)
recruitment(param = dat, tsample = 0.25)


# several samples
data(trout)

# add growth parameters
trout$Linf <- 16
trout$K <- 0.4

# retrieve sampling times from catch matrix
s_dates <- strsplit(colnames(trout$catch),"X")
s_dates <- unlist(lapply(s_dates, function(x) return(x[2])))
s_dates <- as.POSIXlt(s_dates, format="\%d.\%m.\%Y")

recruitment(trout, tsample = s_dates$yday/365)

}
\references{
Moreau, J., & Cuende, F. X., 1991. On improving the resolution of the recruitment
patterns of fishes. \emph{Fishbyte}, 9(1), 45-46.

Pauly, D., 1982. Studying single-species dynamics in a tropical multispecies context.
In Theory and management of tropical fisheries. \emph{ICLARM Conference Proceedings}
(Vol. 9, No. 360, pp. 33-70).

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.

#@export

Moreau, J., & Cuende, F. X., 1991. On improving the resolution of the recruitment
patterns of fishes. \emph{Fishbyte}, 9(1), 45-46.

Pauly, D., 1982. Studying single-species dynamics in a tropical multispecies context.
In Theory and management of tropical fisheries. \emph{ICLARM Conference Proceedings}
(Vol. 9, No. 360, pp. 33-70).

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.

#@export

Brey, T., Soriano, M., Pauly, D., 1988. Electronic length frequency analysis. A
revised and expanded user's guide to ELEFAN 0, 1 and 2. (Second edition).
Berichte aus dem Institut für Meereskunde Kiel, No 177, 31p.

Moreau, J., & Cuende, F. X., 1991. On improving the resolution of the recruitment
patterns of fishes. \emph{Fishbyte}, 9(1), 45-46.

Pauly, D., 1982. Studying single-species dynamics in a tropical multispecies context.
In Theory and management of tropical fisheries. \emph{ICLARM Conference Proceedings}
(Vol. 9, No. 360, pp. 33-70).

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.
}

