% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RikhterEfanovsM.R
\name{RikhterEfanovsM}
\alias{RikhterEfanovsM}
\title{Rikhter and Efanov's formula}
\usage{
RikhterEfanovsM(Tm50)
}
\arguments{
\item{Tm50}{Age when 50\% of the population is mature [year] ("the age of massive maturation")}
}
\description{
Rikhter and Efanov's formula to calculate natural mortality (M)
}
\details{
Rikhter and Efanov's formula to calculate natural mortality (M).
}
\examples{
RikhterEfanovsM(Tm50 = 1)
}

