% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SchaeferFox.R
\name{SchaeferFox}
\alias{SchaeferFox}
\title{Schaefer and Fox models}
\usage{
SchaeferFox(Y, f)
}
\arguments{
\item{Y}{Yield as vector (catch in weight)}

\item{f}{Fishing effort as vector}
}
\description{
Schaefer and Fox models are examples of so called surplus production models, and are used to calculate MSY and BMSY and so on.
}
\details{
MSY
}
\examples{
data(ex.SchaeferFox)
SchaeferFox(Y = ex.SchaeferFox$yield, f = ex.SchaeferFox$effort)
}
\references{
Graham 1935 Schaefer 1954 Fox 1970
}

