% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchaeferFox.R
\name{SchaeferFox}
\alias{SchaeferFox}
\title{Schaefer and Fox models}
\usage{
SchaeferFox(Y, f)
}
\arguments{
\item{Y}{Yield as vector (catch in weight)}

\item{f}{Fishing effort as vector}
}
\description{
Schaefer and Fox models are examples of so called surplus production models, and are used to calculate MSY and BMSY and so on.
}
\details{
MSY
}
\examples{
# load data
data(trawl_fishery_Java)

# run model
output = SchaeferFox(Y = trawl_fishery_Java$yield,
                     f = trawl_fishery_Java$effort)
output[[2]]

}
\references{
Graham 1935 Schaefer 1954 Fox 1970
}

