% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ogive.R
\name{select_ogive}
\alias{select_ogive}
\title{Selectivity function}
\usage{
select_ogive(s_list, classes, Lt, Lc = NA)
}
\arguments{
\item{s_list}{list of selectivity parameters}

\item{classes}{age or size classes}

\item{Lt}{lengths corresponding to age classes}

\item{Lc}{length-at-first-capture Default = \code{NA}}
}
\description{
This function estimates the selecitvity.
   First sentence. second sentence.
}
\details{
To calculate selection factor (SF), L25, L50 and L75 for trawl nets /fisheries.
}
\examples{
# create list with selectivity information
select.list <- list(selecType = 'knife_edge',  #or 'gillnet' or 'trawl_ogive'
   Lc = 34,tc = 5,selecDist = 'lognormal',    #or 'normal_fixed'
   mesh_size = 8.1,mesh_size1 = 9.1,select_p1 = 21.1,select_p2 = 23.8)

# create vector with mid lengths
Lt <- c(12,22,29,34,38,41,43,45,46,47,48,49,49)

# run model
select_ogive(select.list, seq(1,13,1),Lt)

}
\references{
xx
}

