% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startingPoint2tanchor.R
\name{startingPoint2tanchor}
\alias{startingPoint2tanchor}
\title{Convert FiSAT's starting point to t_anchor value}
\usage{
startingPoint2tanchor(param, startingLength, startingSample)
}
\arguments{
\item{param}{lfq object with dates, midLengths, catch, and growth parameters 'Linf'
and 'K' from VBGF}

\item{startingLength}{starting length as returned by FiSAT, indicating the length within
the starting sample cut by a growth curve}

\item{startingSample}{starting sample as returned by FiSAT, indicating the sample which is
cut by a growth curve}
}
\value{
lfq object with t_anchor value
}
\description{
Starting points returned or chosen within FiSAT are not supported
   in TropFishR. Instead \code{t_anchor} takes on the job of anchoring VBGF growth curves on
   a temporal axis. This function allows to convert FiSAT's starting points to \code{t_anchor} values
}
\examples{
data(synLFQ1)
synLFQ1$Linf <- 54
synLFQ1$K <- 0.6
lfqNEW <- startingPoint2tanchor(synLFQ1, startingLength = 20, startingSample = 4)

}
\keyword{function}
\keyword{lfq}
\keyword{startingPoints}
\keyword{t_anchor}

