% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_sim.R
\name{stock_sim}
\alias{stock_sim}
\title{Simulating stock}
\usage{
stock_sim(param, unit.time = "year", stock_size_1 = NA, plus.group = NA)
}
\arguments{
\item{param}{a list consisting of following parameters:
\itemize{
  \item \strong{age} or \strong{midLengths}: midpoints of the length class as vector (length-frequency
  data) or ages as vector (age composition data),
  \item \strong{meanWeight}: mean weight in kg per age class or length group,
  \item \strong{meanValue}: mean value per kg fish per age class or length group,
  \item \strong{FM}: fishing mortality rates per age class or length group,
  \item \strong{M} or \strong{Z}: the natural or total mortality rate;
}}

\item{unit.time}{indicates if the age groups are per month (\code{"month"}) or
per year (\code{"year"}). Default: \code{"year"}}

\item{stock_size_1}{Stock size of smallest age/length group}

\item{plus.group}{indicates age/length group, which should be turned into a plus
group (i.e. all groups above are comprised in one group)}
}
\value{
A list with the input parameters and following list objects:
\itemize{
  \item \strong{dt}: delta t,
  \item \strong{N}: population numbers,
  \item \strong{dead}: number of deaths due to natural mortality,
  \item \strong{C}: catch,
  \item \strong{Y}: yield,
  \item \strong{B}: biomass,
  \item \strong{V}: value,
  \item \strong{totals}: total values:
  \itemize{
  \item \strong{tot.C} total catch,
  \item \strong{tot.Y} total yield,
  \item \strong{tot.V} total value,
  \item \strong{meanB} mean biomass;
  },
}
}
\description{
This function estimates stock size, biomass and the yield of a stock from
   the fishing mortality per age class or length group. It is function is embedded in the
   Thompson and Bell model (prediction model: \code{\link{predict_mod}}).
}
\details{
better to treat last group always as a plus group...
}
\examples{
# age-based stock simulation
# load data
data(shrimps)

# run model
# option 1: without plus group
stock_sim(shrimps)

# option 2: with plus group
stock_sim(shrimps, plus.group = 11)

# length-based stock simulation
# load data
data(hake)

# run model
stock_sim(param = hake, stock_size_1 = 98919.3)

}
\references{
example 1 : Kuwait (Garcia and van zalinge 1982)

Millar, R. B., & Holst, R. (1997). Estimation of gillnet and hook selectivity using
log-linear models. ICES Journal of Marine Science: Journal du Conseil, 54(3):471-477

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\keyword{function}
\keyword{prediction}

