% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synLFQ1.R
\docType{data}
\name{synLFQ1}
\alias{synLFQ1}
\title{Synthetic length frequency data I from Sparre & Venema (1998)}
\format{A list consisting of:
\itemize{
  \item \code{$midLengths} a vector of the mid lengths of the length groups,
  \item \code{$catch_mat} a matrix with the catches for different years.
}}
\source{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment. Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\usage{
data(synLFQ1)
}
\description{
Synthetic length frequency data I. Can be used for the function
   \code{\link{Bhattacharya}}.
}
\examples{

data(synLFQ1)
str(synLFQ1)


}
\keyword{data}
\keyword{dataset}
\keyword{length-frequency}

