% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synLFQ1.R
\docType{data}
\name{synLFQ1}
\alias{synLFQ1}
\title{Synthetic length-frequency data I}
\format{A list consisting of:
\itemize{
  \item \strong{dates} dates of sampling times (class Date),
  \item \strong{midLengths} midpoints of the length classes,
  \item \strong{catch} matrix with catches/counts per length class (row) and sampling date (column).
}}
\source{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
   Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.
}
\usage{
data(synLFQ1)
}
\description{
Synthetic length-frequency data as provided in Sparre & Venema (1998).
   Can be used to apply the function \code{\link[TropFishR]{Bhattacharya}} or
   \code{\link[TropFishR]{ELEFAN}}.
}
\examples{
data(synLFQ1)
str(synLFQ1)
summary(synLFQ1)


}
\keyword{data}
\keyword{dataset}
\keyword{length-frequency}
