% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synLFQ2.R
\docType{data}
\name{synLFQ2}
\alias{synLFQ2}
\title{Synthetic length frequency data II from Sparre & Venema (1998)}
\format{A list consisting of:
\itemize{
  \item \strong{midLengths}: a vector of the mid lengths of the length groups,
  \item \strong{Linf}: infinite length for investigated species in cm [cm],
  \item \strong{K}: growth coefficent for investigated species per year [1/year],
  \item \strong{catch}: a matrix with the catches for different years.
}}
\source{
Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
   Part 1. Manual. FAO Fisheries Technical Paper, (306.1, Rev. 2). 407 p.
}
\usage{
data(synLFQ2)
}
\description{
Synthetic length frequency data II. Can be used for the calculation of Z
   by means of \code{\link{Z_BevertonHolt}}.
}
\examples{

data(synLFQ2)
str(synLFQ2)


}
\keyword{data}
\keyword{length-frequency}

