% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmtool.R
\name{TFSA}
\alias{TFSA}
\title{Traditional Fish Stock Assessment with TropFishR}
\usage{
TFSA(lfq, m, wla, wlb, fmsy)
}
\arguments{
\item{lfq}{lfq data}

\item{m}{natural mortality}

\item{wla}{coefficient from length-weight relationship}

\item{wlb}{exponent from length-weight relationship}
}
\value{
F/F01 estimated by LCCC + length-based VPA + YPR
}
\description{
Estimate F/F01 based on the length converted catch curve, length-based VPA and
  the Thompson and Bell yield per recruit analysis
}
