% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrawlSelect.R
\name{TrawlSelect}
\alias{TrawlSelect}
\title{Trawl net selctivity}
\usage{
TrawlSelect(param)
}
\arguments{
\item{param}{List with parameters: midlengths of size classes, number in net, number in codend and meshsize of codend}
}
\description{
This function estimates the selecitvity of trawl nets.
   First sentence. second sentence.
}
\details{
To calculate selection factor (SF), L25, L50 and L75 for trawl nets /fisheries.
}
\examples{
# load data
data(data_TrawlSelect)

# run model
output <- TrawlSelect(data_TrawlSelect)

# plot results graphically
plot(output)

}
\references{
xx
}

