% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiting.R
\docType{data}
\name{whiting}
\alias{whiting}
\title{Data of North sea whiting \emph{Merlangius merlangus} caught during the period 1974-1980.}
\format{A list consisting of: 1. a vector with age groups, 2. a vector or matrix with catches}
\source{
ICES, 1981. Report of the \emph{Ad hoc} working group on the use of effort data in assessment, Copenhagen, 2-6 March 1981. ICES C.M. 1981/G:5 (mimeo)
}
\usage{
data(whiting)
}
\description{
Data of North sea whiting \emph{Merlangius merlangus} caught during the period 1974-1980.
  Can be used for \code{\link{CatchCurve}} and \code{\link{VPA}}
}
\examples{
data(whiting)
str(whiting)


}
\keyword{catch}
\keyword{dataset}

