% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_BevertonHolt.R
\name{Z_BevertonHolt}
\alias{Z_BevertonHolt}
\title{Beverton and Holt's Z-Equations & Powell and Wetherall method}
\usage{
Z_BevertonHolt(param, catch_column = NA, PowellWetherall = FALSE)
}
\arguments{
\item{param}{a list consisting of following parameters:
\itemize{
  \item \strong{age} or \strong{midLengths}: midpoints of the length class as vector (length-frequency
  data) or ages as vector (age composition data),
  \item \strong{Linf}: infinite length for investigated species in cm [cm],
  \item \strong{K}: growth coefficent for investigated species per year [1/year],
  \item \strong{t0}: theoretical time zero, at which individuals of this species hatch,
  \item \strong{catch}: catch as vector, or a matrix with catches of subsequent years;
}}

\item{catch_column}{optional; in case catch is a matrix or data.frame, a number
indicating which column of the matrix should be analysed (Default: \code{NA}).}

\item{PowellWetherall}{logical; if \code{TRUE} the Powell Wetherall method is
applied (Default: \code{FALSE}).}
}
\description{
A method to calculate Z from a range of equations derived by Beverton and
   Holt (1956) with the option to perform the Powell and Wetherall method (Powell, 1979;
   Wetherall et al., 1987). Refers to Chapter 4.5 in Sparre & Venema (1998).
}
\details{
The first length group or age class within the list object \strong{midLengths} or
   \strong{age} will be used as the Lprim or tprime (length of recruitment to fishery).
   In the case of the Powell and Wetherall method, this function includes a so called
   'locator' function, which asks you to choose two points from a graph manually. The
   two points which you choose by clicking on the plot in the graphical device represent
   the start and end of the data points, which should be used for the analysis. Based
   on these points the regression line is calculated. The Powell and Wetherall method
   only works with length-frequency data.
}
\examples{
#_______________________________________________
# Beverton and Holt's Z method
# based on length data
# load data
data(synLFQ2)

# run model
Z_BevertonHolt(synLFQ2, catch_column = 2)

# based on age data
# load data
data(synCAA1)

#run model
Z_BevertonHolt(synCAA1, catch_column = 3)

\donttest{
#_______________________________________________
# Powell and Wetherall method
# load data
data(synLFQ3)

# run model
Z_BevertonHolt(synLFQ3, PowellWetherall = T)
}

}
\references{
Beverton R.J.H and S.J. Holt, 1956. A review of methods of estimating mortality rates
in exploited fish populations, with special reference to sources of bias in catch
sampling. \emph{Rapp.P.-v.Réun.CIEM}, 140:67-83

Powell, D.G., 1979. Estimation of mortality and growth parameters from the length-
frequency of a catch. \emph{Rapp.P.-v.Réun.CIEM}, 175:167-169

Sparre, P., Venema, S.C., 1998. Introduction to tropical fish stock assessment.
Part 1. Manual. \emph{FAO Fisheries Technical Paper}, (306.1, Rev. 2). 407 p.

Wetherall, J.A., J.J. Polovina and S. Ralston, 1987. Estimating growth and mortality
in steady-state fish stocks from length-frequency data. \emph{ICLARM Conf.Proc.},
(13):53-74
}
\keyword{function}
\keyword{mortality}

