% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Z_CPUE.R
\name{Z_CPUE}
\alias{Z_CPUE}
\title{Methods to calculate Z from CPUE data}
\usage{
Z_CPUE(cohort, age, CPUE, method)
}
\arguments{
\item{cohort}{}

\item{age}{}

\item{CPUE}{}

\item{method}{One of 3 methods to caluclate Z: "standard", "Heincke"}
}
\value{
A matrix with total moratlity rates with column names ... and row names ... .
}
\description{
Method to estimate the total mortality (Z) from CPUE data.
}
\details{
Z from CPUE data
}
\examples{
data("ex.Z_CPUE")
with(ex.Z_CPUE,Z_CPUE(cohort, age.t1, CPUE, method = "standard"))
}
\references{
Jones ???  Sparre?
}

